### 2021-04-09 #3165 @whedon
[PRE REVIEW]: popsynth: A generic astrophysical population synthesis framework
**Submitting author:** @grburgess (<a href="http://orcid.org/0000-0003-3345-9515">J. Michael Burgess</a>)
**Repository:** <a href="https://github.com/grburgess/popsynth" target ="_blank">https://github.com/grburgess/popsynth</a>
**Version:** v1.0.1
