### 2021-04-10 8ce80d1
add extra parameter to sfr dist
