+++ popsynth/populations/bpl_population.py
@@ -139,6 +139,7 @@ class BPLSFRPopulation(SFRPopulation):
     def __init__(
         self,
         r0,
+        a,
         rise,
         decay,
         peak,
@@ -154,6 +155,7 @@ class BPLSFRPopulation(SFRPopulation):
         """FIXME! briefly describe function
 
         :param r0:
+        :param a:
         :param rise:
         :param decay:
         :param peak:
