+++ popsynth/populations/pareto_populations.py
@@ -103,6 +103,7 @@ class ParetoSFRPopulation(SFRPopulation):
     def __init__(
         self,
         r0,
+        a,
         rise,
         decay,
         peak,
@@ -115,6 +116,7 @@ class ParetoSFRPopulation(SFRPopulation):
         """FIXME! briefly describe function
 
         :param r0:
+        :param a:
         :param rise:
         :param decay:
         :param peak:
