+++ popsynth/populations/lognormal_population.py
@@ -242,6 +242,7 @@ class Log10NormalSFRPopulation(SFRPopulation):
     def __init__(
         self,
         r0,
+        a,
         rise,
         decay,
         peak,
@@ -254,6 +255,7 @@ class Log10NormalSFRPopulation(SFRPopulation):
         """FIXME! briefly describe function
 
         :param r0:
+        :param a:
         :param rise:
         :param decay:
         :param peak:
@@ -272,6 +274,7 @@ class Log10NormalSFRPopulation(SFRPopulation):
 
         super(Log10NormalSFRPopulation, self).__init__(
             r0=r0,
+            a=a,
             rise=rise,
             decay=decay,
             peak=peak,
