### 2021-04-11 b64977f
skip selections that are unity
