+++ popsynth/selection_probability/__init__.py
@@ -1,8 +1,9 @@
 from .flux_selectors import HardFluxSelection, SoftFluxSelection
 from .generic_selectors import (BernoulliSelection, BoxSelection, LowerBound,
                                 SoftSelection, UnitySelection, UpperBound)
-from .selection_probability import DummySelection, SelectionProbabilty
+from .selection_probability import (DummySelection, SelectionParameter,
+                                    SelectionProbabilty)
 from .spatial_selection import SpatialSelection
 
 __all__ = ["HardFluxSelection", "SoftFluxSelection", "BernoulliSelection", "LowerBound", "UpperBound",
-                                "SoftSelection", "UnitySelection", "SelectionProbabilty", "SpatialSelection", "BoxSelection", "DummySelection"]
+                                "SoftSelection", "UnitySelection", "SelectionProbabilty", "SpatialSelection", "BoxSelection", "DummySelection", "SelectionProbabilityParameter"]
