### 2021-04-12 aea5360
fixed up constructor and more tests
