+++ README.md
@@ -1,8 +1,7 @@
-[![Codacy Badge](https://api.codacy.com/project/badge/Grade/5d02c9e6f5c540989a615eb1575863e3)](https://app.codacy.com/gh/grburgess/popsynth?utm_source=github.com&utm_medium=referral&utm_content=grburgess/popsynth&utm_campaign=Badge_Grade_Settings)
-[![Codacy Badge](https://api.codacy.com/project/badge/Grade/b3d66930915f4a3a949465abe5311325)](https://app.codacy.com/app/grburgess/popsynth?utm_source=github.com&utm_medium=referral&utm_content=grburgess/popsynth&utm_campaign=Badge_Grade_Dashboard)
+
 ![CI](https://github.com/grburgess/popsynth/workflows/CI/badge.svg?branch=master)
 [![codecov](https://codecov.io/gh/grburgess/popsynth/branch/master/graph/badge.svg)](https://codecov.io/gh/grburgess/popsynth)
-[![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/ambv/black)
+[![Codacy Badge](https://api.codacy.com/project/badge/Grade/5d02c9e6f5c540989a615eb1575863e3)](https://app.codacy.com/gh/grburgess/popsynth?utm_source=github.com&utm_medium=referral&utm_content=grburgess/popsynth&utm_campaign=Badge_Grade_Settings)
 [![Documentation Status](https://readthedocs.org/projects/popsynth/badge/?version=latest)](https://popsynth.readthedocs.io/en/latest/?badge=latest)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3372456.svg)](https://doi.org/10.5281/zenodo.3372456)
 ![PyPI](https://img.shields.io/pypi/v/popsynth)
@@ -14,6 +13,12 @@
 A simple object oreinted  population synthesis framework designed for testing population inference in Stan.
 It has grown passed its orginal intent to be able to produce populations/catalogs from a variety of luminosity functions, redshift distributions, and even more abstractions. Documentation coming soon.
 It also makes a lot of pretty pictures.
+
+## Installation
+```bash
+pip install popsynth
+```
+
 
 Note: **This is not synth pop!** If you were looking for some hard driving beats out of a yamaha keyboard with bells... look elsewhere
 
