+++ popsynth/data/pop.yml
@@ -41,24 +41,24 @@ spatial selection:
 # the are custom!
 
 auxiliary samplers:
-    NormalAuxSampler:
-        name: stellar_mass
+    stellar_mass:
+        type: NormalAuxSampler
         observed: False
         mu: 0
         sigma: 1
         selection:
-        secondary:
+        secondary: []
         init variables:
 
-    DemoSampler:
-        name: demo
+    demo:
+        type: DemoSampler
         observed: False
         selection:
             UpperBound:
                 boundary: 20
 
-    DemoSampler2:
-        name: demo2
+    demo2:
+        type: DemoSampler2
         observed: True
         selection:
         secondary: [demo, stellar_mass] # other samplers that this sampler depends on
