### 2021-04-23 #3165 @warrickball
Hi @xuanxu ! I've just had a look at the paper and a cursory look at the repo and it isn't obviously my area of expertise. While the package can be used to synthesize populations of stars, the framework is quite generic, so the code doesn't appear to involve any stellar astrophysics.

I'm not sure who the other potential reviewers are but I'd recommend someone who is better versed in the statistics of astronomical surveys.

If you have trouble finding another reviewer, do ask again, but I disclaim that I'll be approaching with limited knowledge of the science case at hand.
