### 2021-04-30 #3165 @ilyamandel
Alas, I can't help. :(
I just started as an editor at ApJ earlier this year, so I am fairly busy with that at the moment.
I can recommend Jeff Andrews or Heloise Stevance.
