### 2021-05-01 582ba86
Store is_rate for cosmological spatial distributions
