+++ popsynth/population_synth.py
@@ -10,20 +10,28 @@ import scipy.special as sf
 import scipy.stats as stats
 import yaml
 from IPython.display import Markdown, Math, display
+
 # from numpy.typing import np.ndarray
 from numba import float64, jit, njit, prange
 
 from popsynth.auxiliary_sampler import AuxiliarySampler, DerivedLumAuxSampler
 from popsynth.distribution import LuminosityDistribution, SpatialDistribution
+from popsynth.distributions.cosmological_distribution import CosmologicalDistribution
 from popsynth.population import Population
-from popsynth.selection_probability import (BernoulliSelection,
-                                            HardFluxSelection,
-                                            SelectionProbabilty,
-                                            SoftFluxSelection, UnitySelection)
+from popsynth.selection_probability import (
+    BernoulliSelection,
+    HardFluxSelection,
+    SelectionProbabilty,
+    SoftFluxSelection,
+    UnitySelection,
+)
 from popsynth.utils.logging import setup_logger
 from popsynth.utils.progress_bar import progress_bar
-from popsynth.utils.registry import (auxiliary_parameter_registry,
-                                     distribution_registry, selection_registry)
+from popsynth.utils.registry import (
+    auxiliary_parameter_registry,
+    distribution_registry,
+    selection_registry,
+)
 
 log = setup_logger(__name__)
 
@@ -58,9 +66,7 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
         self._graph = nx.DiGraph()  # type: nx.Digraph
 
-        if not isinstance(
-                spatial_distribution,
-                SpatialDistribution):
+        if not isinstance(spatial_distribution, SpatialDistribution):
 
             log.error("the spatial_distribution is the wrong type")
 
@@ -70,9 +76,7 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
         if luminosity_distribution is not None:
 
-            if not isinstance(
-                luminosity_distribution, LuminosityDistribution
-            ):
+            if not isinstance(luminosity_distribution, LuminosityDistribution):
 
                 log.error("the luminosity_distribution is the wrong type")
 
@@ -87,13 +91,16 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
         self._has_derived_luminosity = False  # type: bool
         self._derived_luminosity_sampler = (
-            None)  # type: Union[DerivedLumAuxSampler, None]
+            None
+        )  # type: Union[DerivedLumAuxSampler, None]
 
         # set the selections be fully seen unless it is set by the user
         self._distance_selector: SelectionProbabilty = UnitySelection(
-            name="unity distance selector")
+            name="unity distance selector"
+        )
         self._flux_selector: SelectionProbabilty = UnitySelection(
-            name="unity flux selector")
+            name="unity flux selector"
+        )
 
         # check to see if the selectors are set
         self._distance_selector_set: bool = False
@@ -119,7 +126,7 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
     def clean(self, reset: bool = False):
         """
-        clean the auxiliary samplers, selections, etc 
+        clean the auxiliary samplers, selections, etc
         from the population synth
         :param reset: reset any attached distributions and samplers
 
@@ -163,8 +170,7 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
             self._distance_selector.reset()
 
-        self._distance_selector = UnitySelection(
-            name="unity distance selector")
+        self._distance_selector = UnitySelection(name="unity distance selector")
 
         log.warning("removing spatial selector")
 
@@ -184,8 +190,12 @@ class PopulationSynth(object, metaclass=ABCMeta):
         """
         with open(file_name, "w") as f:
 
-            yaml.dump(stream=f, data=self.to_dict(),
[…]
