### 2021-05-01 2e257e9
Read is_rate from dict when loading popsynth
