+++ popsynth/aux_samplers/__init__.py
@@ -3,10 +3,7 @@ from .lognormal_aux_sampler import LogNormalAuxSampler, Log10NormalAuxSampler
 from .normal_aux_sampler import NormalAuxSampler
 from .trunc_normal_aux_sampler import TruncatedNormalAuxSampler
 
-__all__ = ["ViewingAngleSampler",
-           "LogNormalAuxSampler",
-           "Log10NormalAuxSampler",
-           "NormalAuxSampler",
-           "TruncatedNormalAuxSampler"
-           
-           ]
+__all__ = [
+    "ViewingAngleSampler", "LogNormalAuxSampler", "Log10NormalAuxSampler",
+    "NormalAuxSampler", "TruncatedNormalAuxSampler"
+]
