+++ popsynth/population_synth.py
@@ -91,16 +91,13 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
         self._has_derived_luminosity = False  # type: bool
         self._derived_luminosity_sampler = (
-            None
-        )  # type: Union[DerivedLumAuxSampler, None]
+            None)  # type: Union[DerivedLumAuxSampler, None]
 
         # set the selections be fully seen unless it is set by the user
         self._distance_selector: SelectionProbabilty = UnitySelection(
-            name="unity distance selector"
-        )
+            name="unity distance selector")
         self._flux_selector: SelectionProbabilty = UnitySelection(
-            name="unity flux selector"
-        )
+            name="unity flux selector")
 
         # check to see if the selectors are set
         self._distance_selector_set: bool = False
@@ -170,7 +167,8 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
             self._distance_selector.reset()
 
-        self._distance_selector = UnitySelection(name="unity distance selector")
+        self._distance_selector = UnitySelection(
+            name="unity distance selector")
 
         log.warning("removing spatial selector")
 
@@ -211,14 +209,15 @@ class PopulationSynth(object, metaclass=ABCMeta):
         spatial_distribution = {}
 
         spatial_distribution[
-            self._spatial_distribution._distribution_name
-        ] = self._spatial_distribution.truth
+            self._spatial_distribution.
+            _distribution_name] = self._spatial_distribution.truth
 
         # store is_rate if cosmological distribution
 
         if isinstance(spatial_distribution, CosmologicalDistribution):
 
-            spatial_distribution["is_rate"] = self._spatial_distribution._is_rate
+            spatial_distribution[
+                "is_rate"] = self._spatial_distribution._is_rate
 
         output["spatial distribution"] = spatial_distribution
 
@@ -230,8 +229,8 @@ class PopulationSynth(object, metaclass=ABCMeta):
             luminosity_distribution = {}
 
             luminosity_distribution[
-                self._luminosity_distribution._distribution_name
-            ] = self._luminosity_distribution.truth
+                self._luminosity_distribution.
+                _distribution_name] = self._luminosity_distribution.truth
 
             output["luminosity distribution"] = luminosity_distribution
 
@@ -239,9 +238,8 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
             flux_selection = {}
 
-            flux_selection[
-                self._flux_selector._selection_name
-            ] = self._flux_selector.parameters
+            flux_selection[self._flux_selector.
+                           _selection_name] = self._flux_selector.parameters
 
             output["flux selection"] = flux_selection
 
@@ -250,8 +248,8 @@ class PopulationSynth(object, metaclass=ABCMeta):
             distance_selection = {}
 
             distance_selection[
-                self._distance_selector._selection_name
-            ] = self._distance_selector.parameters
+                self._distance_selector.
+                _selection_name] = self._distance_selector.parameters
 
             output["distance selection"] = distance_selection
 
@@ -260,8 +258,8 @@ class PopulationSynth(object, metaclass=ABCMeta):
             spatial_selection = {}
 
             spatial_selection[
-                self._spatial_selector._selection_name
-            ] = self._spatial_selector.parameters
+                self._spatial_selector.
+                _selection_name] = self._spatial_selector.parameters
 
             output["spatial selection"] = spatial_selection
 
@@ -316,8 +314,7 @@ class PopulationSynth(object, metaclass=ABCMeta):
             # create the instance
 
             luminosity_distribtuion: LuminosityDistribution = distribution_registry[
[…]
