### 2021-05-01 9521964
Add is_rate into test yaml pop files
