### 2021-05-03 b5b2692
revert to numpy arraylike
