### 2021-05-12 #3257 @warrickball
Hi,

I've just made a first pass at reviewing the repository. I'm not an expert on the scientific application but I'm at least an astronomer/astrophysicist and will comment on the repo from that perspective.

First, the code itself seems reasonable overall. I had no problem installing and running it, though the nice-looking 3D plots didn't work in a script or IPython console. (I don't personally use Jupyter notebooks much.) The code certainly appears to represent significant and useful work. As noted in the paper, other population synthesis codes tend to focus on particular kinds of astrophysical object, whereas `popsynth` is clearly meant to be abstract enough to apply to many different types of object.

---

My principal concern is the documentation.  For a start, it appears incomplete.  The API docs linked from the left sidebar are empty and there are two links to the same page (*API* and *popsynth*).  I also came across a `FIXME` in the docstring of `popsynth.ZPowerSphericalPopulation`, which it turns out is in docstrings of many of the spatial populations.  The experimental YAML documentation in the *Quick start* also has a sentence

> We can load this yaml file into a population synth like this:

which abruptly leads to the next section.  (I wouldn't recommend documenting experimental features in the *Quick start*.)

The *Quick start* plunges into an abstract and jargon-laden example. So, though it works and the abstraction makes the code usefully versatile, it isn't at all obvious to me how it applies to a real-world problem, though perhaps this would be obvious to someone in the field.  I think the docs would be better served by relating its parameters to a realistic problem. e.g.  Could the population in the *Quick start* be thought of as AGN?  Similarly, few of the distributions are familiar and I've only seen the equations in the docs produced by the population generator objects' `display()` function (which, like the 3D plots, didn't work for me in scripts or IPython).  It'd be good to have a reference for each equation but this would in effect be done by populating the distributions' docstrings.

On a related note, there are some technical details about what the code aims to do that I don't understand.  Are we trying to create a population of *events* or of *sources*?  I presumed the code was for *sources* but the text of the paper mentions gamma-ray bursts (GRBs) as examples.  I would've thought GRBs are one-off events: once a GRB has been detected from a source, we wouldn't expect to observe another GRB from that source, would we?  If I've understood that correctly, it then isn't obvious to me how `popsynth` simulates a population of GRBs.  Perhaps the "source" of a GRB is really its host galaxy?  The animated spinning figure on the frontpage of the documentation also gives the impression of simulating a population of events.

I think this would be cleared up if there were some mathematical definitions for what exactly is meant by terms like *population*,
*luminosity function*, etc.  I don't expect these to be prominent: they can go in a glossary or something (though they should perhaps go in the paper too).  But without them, it's difficult to know exactly what the code is intended to do.

It's a simple thing but for a non-expert like me, I'd find it easier to understand the docs if the various classes and functions were
internally linked (e.g. with ReST `:py:class:...` directives).

To close this section, you may take my scientific remarks with a pinch of salt.  I'm an not an expert and perhaps your target user knows enough that my comments are moot.

---

The requirements for testing aren't specified anywhere.  I cloned the repo and tried `pytest` in the directory above the `tests` I found, which revealed that `hypothesis` is required.  With that installed I got an error

```
popsynth/test/test_selections.py:7: in <module>
    class GalacticPlaceSelection(popsynth.SpatialSelection):
popsynth/test/test_selections.py:11: in GalacticPlaceSelection
    b_limit = popsynth.SelectionParameter(vmin=0, vmax=90)
E   AttributeError: module 'popsynth' has no attribute 'SelectionParameter'
```

and a bunch of warnings, mostly about "traits".  The CI appears to be passing, though, so perhaps I've missed a dependency or run `pytest` in the wrong place.  My platform information (from `pytest`) is:

```
rootdir: ..., configfile: setup.cfg, testpaths: popsynth/test
plugins: cov-2.10.1, remotedata-0.3.2, hypothesis-6.12.0
```

---

Finally, two small points.  First, there are no "clear guidelines for third parties wishing to 1) Contribute to the software 2) Report
issues or problems with the software 3) Seek support".  All of this can be done through GitHub but the repo or docs should say so.  (On a related note, the docs don't seem to link to the GitHub repo anywhere.)

Second, the `SkyPy` and `firesong` references in the paper render as

> al., A. A. et. (2021). SkyPy: A package for modelling the universe. InGitHub repository. GitHub.https://github.com/skypyproject/skypy
> al., I. T. et. (2021). FIRst extragalactic simulation of neutrino and gamma-ray. InGitHub repository. GitHub.https://github.com/icecube/FIRESONG

suggesting an error in the author lists but these references can be replaced.  `firesong` has a [JOSS paper](https://joss.theoj.org/papers/10.21105/joss.03194) and `SkyPy` has a [citation file](https://github.com/skypyproject/skypy/blob/main/CITATION.rst) that points to recommended DOIs and BibTeX entries on [Zenodo](https://zenodo.org/record/3755531).
