### 2021-05-12 #3257 @grburgess
Thank you! I will begin to implement these changes, but here are some quick responses:

> My principal concern is the documentation. For a start, it appears incomplete. The API docs linked from the left sidebar are empty and there are two links to the same page (API and popsynth). I also came across a FIXME in the docstring of popsynth.ZPowerSphericalPopulation, which it turns out is in docstrings of many of the spatial populations.

Yes, I need to better doc all the included distributions. Many of these are things that I use and it is not meant to have an exhaustive list because users can easily add their own. If needed, I can add others, but others as well as myself have slowly been adding on things in other repos as they are needed.

> The experimental YAML documentation in the Quick start also has a sentence We can load this yaml file into a population synth like this: which abruptly leads to the next section. (I wouldn't recommend documenting experimental features in the Quick start.)

Oops. This seems like a commit that got behind. Will fix.

> The Quick start plunges into an abstract and jargon-laden example. So, though it works and the abstraction makes the code usefully versatile, it isn't at all obvious to me how it applies to a real-world problem, though perhaps this would be obvious to someone in the field. I think the docs would be better served by relating its parameters to a realistic problem. e.g. Could the population in the Quick start be thought of as AGN? Similarly, few of the distributions are familiar and I've only seen the equations in the docs produced by the population generator objects' display() function (which, like the 3D plots, didn't work for me in scripts or IPython). It'd be good to have a reference for each equation but this would in effect be done by populating the distributions' docstrings.

This is a good point. There were originally equations but looking at other JOSS papers I thought this was too much. I'll gladly add these back in. Sorry about the iPython part, I'm using the Ipython html display capabilities for this stuff, and it is really only meant to be eye-candy and double checking when examining the functions. I think you are right that a more concrete example or changing "population" to "AGN" or "GRB" would make the connection to the abstraction a little easier. I will also add some specific examples.

> On a related note, there are some technical details about what the code aims to do that I don't understand. Are we trying to create a population of events or of sources? I presumed the code was for sources but the text of the paper mentions gamma-ray bursts (GRBs) as examples. I would've thought GRBs are one-off events: once a GRB has been detected from a source, we wouldn't expect to observe another GRB from that source, would we? If I've understood that correctly, it then isn't obvious to me how popsynth simulates a population of GRBs. Perhaps the "source" of a GRB is really its host galaxy? The animated spinning figure on the frontpage of the documentation also gives the impression of simulating a population of events.

Yes, this is confusing. The idea original idea  was to create a "survey" or an observed population of fluxes sampled from their observation likelihoods. As the project grew and others were using it, the purpose became more abstract because the structure lets you easily generate linked/dependent quantities. I think adding the math behind the idea will clear some of this up. The code was originally intended to test hierarchal Bayesian models that fit a population of observations, but it has grown..

> I think this would be cleared up if there were some mathematical definitions for what exactly is meant by terms like population,
luminosity function, etc. I don't expect these to be prominent: they can go in a glossary or something (though they should perhaps go in the paper too). But without them, it's difficult to know exactly what the code is intended to do.

Agreed. I will add this back.

> It's a simple thing but for a non-expert like me, I'd find it easier to understand the docs if the various classes and functions were
internally linked (e.g. with ReST :py:class:... directives).

I agree, I was trying to use auto doc to generate everything in real time with GitHub actions, but this seems to be failing sporadically. I will try and implement your suggestion.

> To close this section, you may take my scientific remarks with a pinch of salt. I'm an not an expert and perhaps your target user knows enough that my comments are moot.

Ah ok, seems there might be an issue in the setup.cfg. I may have beat the CI environment into doing this automatically so I will try and get that working locally. I'm actually not a fan of installing hypotheses if I do not need it... but I will add that to the reqs.

> Finally, two small points. First, there are no "clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support". All of this can be done through GitHub but the repo or docs should say so. (On a related note, the docs don't seem to link to the GitHub repo anywhere.)

Will fix

> Second, the SkyPy and firesong references in the paper render as

Will fix this, I see the Firesong JOSS paper was just out today.
