### 2021-05-24 #3257 @HeloiseS
Hi !

just wanted to check in to say that I never got the notification for this (even on github) I should have sent an email earlier to check everything was working!

I'll be taking a look at this in the next few days but FYI I'm in the middle of starting fellowship applications so I'm doing my best, not just dragging my feet - looking forward to looking at this!

Cheers,
Heloise
