### 2021-06-01 #3257 @HeloiseS
Hi,

I think my comments will echo what the other referee said. First of all congratulations on your code-base! As far as I can see it's very clean and maintained to a high standard of testing.

## General comments

My first and main comment is that after read the paper and gone through the tutorials, I am still not sure what use cases I would apply this code to - I think that the level of abstraction and the heavy use of jargon (that sometimes has multiple meaning across different fields) is a big part of the problem.

What would help would be some real life examples: The other referee mentioned GRBs; my first thought was also transients, both galactic and extra galactic ones. On reading the paper I was imagining that I could give `popsynth` transient rates (you do mention those at the beginning of the Procedure section) derived from stellar population synthesis codes, add observational biases, take into account my telescope's magnitude limits in specific bands, include an approximation of the milky way disc extinction, add a telescope latitude, a program time (say 3 nights) and BAM! get an estimate of how many hits I can expect.

Having looked at `popsynth`, I have no idea where to start to make realistic observations/simulations. I would encourage you to make sure your tutorials are clear even to Master's students/ early PhD students. If your code is open source and in JOSS it is intended for the whole community and at the moment the learning curve is way too high for your really nice code to be used by a large population.

Are there use cases out there already demonstrating how this general and abstract API can be used to address specific astronomy problems? You should cite those in the paper :)

Finally I suggest you do a thorough proof-read of the documentation I came across a few typos.

## Validating and referencing the science + Addressing Jargon

###  **Distributions**
What are the distributions? What do they do?
I understand the concept of the spatial distribution, I am not sure what the luminosity function actually is and where it comes from? Is it source specific? Why does it differ from a basic inverse square law (I assume a majority of astro people know their inverse square law so that can be a good start to anchor your explanation in knowledge that most users will share). I think this goes back to adding more explanation so that a wide audience understands the science behind the code and also adding references as mentioned by reviewer 1.

Also I am concerned that I don't see any validation: How do I know these give me realistic simulations? I liked your graphs showing the "selected" and "non-selected" objects - I'd like to see how that compares to other theoretical predictions. Say the feed the rate of supernovae from stellar evolution codes into `popsynth` and see if you roughly retrieve the number of observed supernovae (given a volume).

### **Auxiliary Samplers**

> "This allows you to build up arbitrailiy complex dependencies between parameters which can lead to diverse populations."

=> But why? I need a solid reason to add complexity.

Also I don't understand the justification for the currently implemented Auxiliary Samplers: I do understand the `ViewingAngle`, but what's the science behind having a log-normal or a log10 Auxiliary Sampler?

**Jargon**

Here are a couple of examples:
What is a SFR-like redshift? What is a Schecter luminosity function? What populations are we synthesysing? Stars? Transients? Observations? All three?

## Jupyter notebooks

None of the 3D graphs render for me.

### **Dependencies**
I installed `popsynth` in a fresh conda env and I was missing a few dependencies to run the jupyter notebooks
- `jupyterthemes`
- `pygraphviz`

That last one straight up failed to install. I am seeing [this issue in their GitHub](https://github.com/pygraphviz/pygraphviz/issues/163) indicating it could be missing some development files.

That is a lot of complex dependency for flowchart and I don't know if it adds much. I would suggest reconsidering which dependencies are absolutely necessary for the good running of the code and removing the rest. If you want to illustrate class inheritance of specific cases, make a nice graph and add it as a picture. The user will not be making these graphs so they don't need to have to carry its dependency.

###  **YAML**
Like the other referee I have reservations about putting experimental material in the quick start. The other problem is that it doesn't run for me. I get the following error:
`FileNotFoundError: [Errno 2] No such file or directory: '/home/fste075/anaconda3/envs/test_popsynth/lib/python3.9/site-packages/popsynth/data/pop.yml'`

This is a probably a result of the fact that I have it all in conda, but a fair few users will have your code running in a specific conda env so it's worth thinking about.

###  **BernouilliSelection**

I got the following error:
`FileNotFoundError: [Errno 2] No such file or directory: '/home/fste075/anaconda3/envs/test_popsynth/lib/python3.9/site-packages/popsynth/data/pop.yml'`

### Other

- a lot of your docstrings are incomplete. Because of that, most of your documentatoin under `popsynth` is empty headings.
- add comments with the science references in the code so other users know where your equations come from
