### 2021-06-09 2d93c1d
Add simple delta distribution
