### 2021-06-11 5b6ec83
Add delta_aux_sampler to easily switch out in testing
