+++ popsynth/aux_samplers/delta_aux_sampler.py
@@ -33,9 +33,9 @@ class DeltaAuxSampler(AuxiliarySampler):
 
         if self._is_observed:
 
-            self._obs_values = stats.norm.rvs(
-                loc=self._true_values, scale=self.sigma, size=size
-            )
+            self._obs_values = stats.norm.rvs(loc=self._true_values,
+                                              scale=self.sigma,
+                                              size=size)
 
         else:
 
