+++ popsynth/test/test_selections.py
@@ -18,11 +18,13 @@ class GalacticPlaceSelection(popsynth.SpatialSelection):
 
     def draw(self, size: int):
 
-        g_coor = SkyCoord(self._spatial_distribution.ra, self._spatial_distribution.dec, unit="deg",
+        g_coor = SkyCoord(self._spatial_distribution.ra,
+                          self._spatial_distribution.dec,
+                          unit="deg",
                           frame="icrs").transform_to("galactic")
 
-        self._selection = (g_coor.b.deg >= self.b_limit) | (
-            g_coor.b.deg <= -self.b_limit)
+        self._selection = (g_coor.b.deg >= self.b_limit) | (g_coor.b.deg <=
+                                                            -self.b_limit)
 
 
 def test_spatial_selection():
