+++ setup.cfg
@@ -50,7 +50,7 @@ install_requires =
 tests_require =
     pytest
     pytest-codecov
-
+    hypothesis
 
 [tool:pytest]
 # Options for py.test:
