+++ popsynth/test/test_distributions.py
@@ -1,9 +1,9 @@
 import numpy as np
 
 from popsynth.distribution import LuminosityDistribution, SpatialDistribution
+from popsynth.distributions.delta_distribution import DeltaDistribution
 from popsynth.distributions.flatland_distribution import FlatlandDistribution
-from popsynth.distributions.spiral_galaxy_distribution import \
-    SpiralGalaxyDistribution
+from popsynth.distributions.spiral_galaxy_distribution import SpiralGalaxyDistribution
 from popsynth.population_synth import PopulationSynth
 
 
@@ -16,9 +16,7 @@ class DummySDistribution(SpatialDistribution):
         # we do not need a "truth" dict here because
         # there are no parameters
 
-        super(DummySDistribution, self).__init__(seed=seed,
-                                                 name="dummy1",
-                                                 form=form)
+        super(DummySDistribution, self).__init__(seed=seed, name="dummy1", form=form)
 
     def differential_volume(self, r):
 
@@ -40,9 +38,7 @@ class DummyLDistribution(LuminosityDistribution):
         # the latex formula for the ditribution
         lf_form = r"1"
 
-        super(DummyLDistribution, self).__init__(seed=seed,
-                                                 name="pareto",
-                                                 form=lf_form)
+        super(DummyLDistribution, self).__init__(seed=seed, name="pareto", form=lf_form)
 
     def phi(self, L):
 
@@ -88,6 +84,24 @@ class MyFlatPopulation(PopulationSynth):
         )
 
 
+class MyDeltaPopulation(PopulationSynth):
+    def __init__(self, r_max=5, seed=1234):
+
+        # instantiate the distributions
+        luminosity_distribution = DeltaDistribution()
+        luminosity_distribution.Lp = 1
+
+        spatial_distribution = FlatlandDistribution()
+        spatial_distribution.Lambda = 1
+        spatial_distribution.r_max = r_max
+
+        super(MyDeltaPopulation, self).__init__(
+            spatial_distribution=spatial_distribution,
+            luminosity_distribution=luminosity_distribution,
+            seed=seed,
+        )
+
+
 def test_distribution_with_no_parameters():
 
     popgen = MyPopulation()
@@ -114,3 +128,10 @@ def test_spiral():
     ld.Lmin = 1
     synth = MWRadialPopulation(rho=1, luminosity_distribution=ld)
     population = synth.draw_survey()
+
+
+def test_delta():
+
+    popgen = MyDeltaPopulation()
+
+    popgen.draw_survey()
