### 2021-06-15 75d6a8f
Test for delta_aux_sampler
