### 2021-06-15 c42671f
Fix typo in GalacticPlaneSelection
