### 2021-06-15 801feab
Test populations with is_rate=False
