+++ popsynth/test/test_distributions.py
@@ -16,7 +16,9 @@ class DummySDistribution(SpatialDistribution):
         # we do not need a "truth" dict here because
         # there are no parameters
 
-        super(DummySDistribution, self).__init__(seed=seed, name="dummy1", form=form)
+        super(DummySDistribution, self).__init__(seed=seed,
+                                                 name="dummy1",
+                                                 form=form)
 
     def differential_volume(self, r):
 
@@ -38,7 +40,9 @@ class DummyLDistribution(LuminosityDistribution):
         # the latex formula for the ditribution
         lf_form = r"1"
 
-        super(DummyLDistribution, self).__init__(seed=seed, name="pareto", form=lf_form)
+        super(DummyLDistribution, self).__init__(seed=seed,
+                                                 name="pareto",
+                                                 form=lf_form)
 
     def phi(self, L):
 
