+++ popsynth/population.py
@@ -23,40 +23,37 @@ from popsynth.utils.spherical_geometry import xyz
 
 log = setup_logger(__name__)
 
-
 wine = "#8F2727"
 dark, dark_highlight, mid, mid_highlight, light, light_highlight = betagen(
     wine)
 
 
 class Population(object):
-    def __init__(
-        self,
-        luminosities: ArrayLike,
-        distances: ArrayLike,
-        known_distances: ArrayLike,
-        known_distance_idx: ArrayLike,
-        unknown_distance_idx: ArrayLike,
-        fluxes: ArrayLike,
-        flux_obs: ArrayLike,
-        selection: ArrayLike,
-        flux_sigma: float,
-        r_max: float,
-        n_model: int,
-        lf_params: Dict[str, Any],
-        spatial_params: Optional[Dict[str, Any]] = None,
-        model_spaces: Optional[ArrayLike] = None,
-        seed: int = 1234,
-        name: Optional[str] = None,
-        spatial_form: Optional[Dict[str, Any]] = None,
-        lf_form: Optional[Dict[str, Any]] = None,
-        auxiliary_quantities: Optional[Dict[str, Any]] = None,
-        truth: Dict[str, float] = {},
-        graph: Optional[Dict[str, Any]] = None,
-        theta=None,
-        phi=None,
-        pop_synth: Optional[Dict[str, Any]] = None
-    ) -> None:
+    def __init__(self,
+                 luminosities: ArrayLike,
+                 distances: ArrayLike,
+                 known_distances: ArrayLike,
+                 known_distance_idx: ArrayLike,
+                 unknown_distance_idx: ArrayLike,
+                 fluxes: ArrayLike,
+                 flux_obs: ArrayLike,
+                 selection: ArrayLike,
+                 flux_sigma: float,
+                 r_max: float,
+                 n_model: int,
+                 lf_params: Dict[str, Any],
+                 spatial_params: Optional[Dict[str, Any]] = None,
+                 model_spaces: Optional[ArrayLike] = None,
+                 seed: int = 1234,
+                 name: Optional[str] = None,
+                 spatial_form: Optional[Dict[str, Any]] = None,
+                 lf_form: Optional[Dict[str, Any]] = None,
+                 auxiliary_quantities: Optional[Dict[str, Any]] = None,
+                 truth: Dict[str, float] = {},
+                 graph: Optional[Dict[str, Any]] = None,
+                 theta=None,
+                 phi=None,
+                 pop_synth: Optional[Dict[str, Any]] = None) -> None:
         """
         A population containing all the simulated variables
 
@@ -625,32 +622,30 @@ class Population(object):
 
         pop_synth = recursively_load_dict_contents_from_group(f, "popsynth")
 
-        return cls(
-            luminosities=luminosities,
-            distances=distances,
-            known_distances=known_distances,
-            known_distance_idx=known_distance_idx,
-            unknown_distance_idx=unknown_distance_idx,
-            fluxes=fluxes,
-            flux_obs=flux_obs,
-            selection=selection,
-            flux_sigma=flux_sigma,
-            n_model=n_model,
-            r_max=r_max,
-            lf_params=lf_params,
-            spatial_params=spatial_params,
-            model_spaces=model_spaces,
-            seed=seed,
-            name=name,
-            spatial_form=spatial_form,
-            lf_form=lf_form,
-            auxiliary_quantities=auxiliary_quantities,
-            truth=truth,
-            graph=graph,
-            theta=theta,
-            phi=phi,
-            pop_synth=pop_synth
-        )
+        return cls(luminosities=luminosities,
+                   distances=distances,
+                   known_distances=known_distances,
[…]
