+++ docs/md/installation.md
@@ -1,0 +1,40 @@
+---
+jupyter:
+  jupytext:
+    formats: ipynb,md
+    text_representation:
+      extension: .md
+      format_name: markdown
+      format_version: '1.3'
+      jupytext_version: 1.11.1
+  kernelspec:
+    display_name: Python 3
+    language: python
+    name: python3
+---
+
+<!-- #region -->
+# Installation
+
+```popsynth``` can be installed with
+
+```bash
+pip install popsynth
+```
+Alternatively, one can install via git
+
+```bash
+git clone https://github.com/grburgess/popsynth
+cd popsynth
+python setup.py install
+```
+
+
+In order to produce graphs of populations, one needs the optional [graphiz](https://graphviz.readthedocs.io/en/stable/) library.
+The 3D plots seen in the documentation require properly setting up the [ipyvolume](https://ipyvolume.readthedocs.io/en/latest/) package and configuring it for your local setup. 
+
+<!-- #endregion -->
+
+```python
+
+```
