### 2021-06-21 9a5aec7
getting numpy doc to work
