+++ docs/index.rst
@@ -29,7 +29,7 @@ selection functions.
 
 
 .. toctree::
-    :maxdepth: 5
+    :maxdepth: 2
     :hidden:
 
     notebooks/installation.ipynb
