+++ docs/requirements.txt
@@ -4,7 +4,7 @@ nbsphinx
 ipykernel
 cython
 jupyterthemes
-numpy
+numpy>=1.20
 scipy
 ipython
 ipyvolume
