+++ .github/workflows/docs.yml
@@ -11,7 +11,7 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v2
         with:
-          python-version: 3.8
+          python-version: 3.9
       
       - name: Setup Graphviz
         uses: ts-graphviz/setup-graphviz@v1
