+++ popsynth/distributions/log_normal_distribution.py
@@ -12,6 +12,16 @@ class LogNormalDistribution(LuminosityDistribution):
 
     def __init__(self, seed=1234, name="lognorm"):
 
+        """
+        a log-normal luminosity distribution
+
+        :param seed: 
+        :type seed: 
+        :param name: 
+        :type name: 
+        :returns: 
+
+        """
         lf_form = r"\frac{\alpha L_{\rm min}^{\alpha}}{L^{\alpha+1}}"
 
         super(LogNormalDistribution, self).__init__(name=name,
