+++ docs/md/quickstart.md
@@ -15,7 +15,7 @@ jupyter:
 
 # Quick start
 
-A simple example of simulating a population via the built-in populations provided.
+A simple example of simulating a population survey via the built-in populations provided.
 
 ```python
 %matplotlib inline
@@ -39,15 +39,70 @@ import warnings
 warnings.simplefilter("ignore")
 ```
 
+```python nbsphinx="hidden"
+class DemoSampler(popsynth.AuxiliarySampler):
+    _auxiliary_sampler_name = "DemoSampler"
+    mu = popsynth.auxiliary_sampler.AuxiliaryParameter(default=2)
+    tau = popsynth.auxiliary_sampler.AuxiliaryParameter(default=1, vmin=0)
+
+    def __init__(self):
+
+        super(DemoSampler, self).__init__("demo", observed=False)
+
+    def true_sampler(self, size):
+
+        self._true_values = np.random.normal(self.mu, self.tau, size=size)
+
+
+class DemoSampler2(popsynth.DerivedLumAuxSampler):
+    _auxiliary_sampler_name = "DemoSampler2"
+    mu = popsynth.auxiliary_sampler.AuxiliaryParameter(default=2)
+    tau = popsynth.auxiliary_sampler.AuxiliaryParameter(default=1, vmin=0)
+    sigma = popsynth.auxiliary_sampler.AuxiliaryParameter(default=1, vmin=0)
+
+    def __init__(self):
+
+        super(DemoSampler2, self).__init__("demo2")
+
+    def true_sampler(self, size):
+
+        secondary = self._secondary_samplers["demo"]
+
+        self._true_values = (
+            (np.random.normal(self.mu, self.tau, size=size))
+            + secondary.true_values
+            - np.log10(1 + self._distance)
+        )
+
+    def observation_sampler(self, size):
+
+        self._obs_values = self._true_values + np.random.normal(
+            0, self.sigma, size=size
+        )
+
+    def compute_luminosity(self):
+
+        secondary = self._secondary_samplers["demo"]
+
+        return (10 ** (self._true_values + 54)) / secondary.true_values
+```
+
+
+
 ## A spherically homogenous population with a pareto luminosity function
 
-**popsynth** comes with several types of populations included, though you can easily construct your own. To create a population synthesizer, one simply instantiates the population form the **popsynth.populations** module.
+**popsynth** comes with several types of populations included, though
+you can easily construct your own. To create a population synthesizer,
+one simply instantiates the population form the
+**popsynth.populations** module. Here, we will simulate a survey that
+has a homogenous spherical spatial distribution and a pareto
+distributed luminosity.
 
 ```python
-homo_pareto_synth = popsynth.populations.ParetoHomogeneousSphericalPopulation(
+homogeneous_pareto_synth = popsynth.populations.ParetoHomogeneousSphericalPopulation(
     Lambda=5, Lmin=1, alpha=2.0  # the density normalization  # lower bound on the LF
 )  # index of the LF
-homo_pareto_synth.display()
+homogeneous_pareto_synth.display()
 ```
 
 
@@ -57,9 +112,9 @@ homo_pareto_synth.display()
 
 options = {"node_color": purple, "node_size": 3000, "width": 0.5}
 
-pos = nx.drawing.nx_agraph.graphviz_layout(homo_pareto_synth.graph, prog="dot")
+pos = nx.drawing.nx_agraph.graphviz_layout(homogeneous_pareto_synth.graph, prog="dot")
 
-nx.draw(homo_pareto_synth.graph, with_labels=True, pos=pos, **options)
+nx.draw(homogeneous_pareto_synth.graph, with_labels=True, pos=pos, **options)
 ```
 
 
[…]
