+++ docs/md/contribute.md
@@ -19,14 +19,24 @@ Contributions to ```popsynth``` are always welcome. They can come in the form of
 
 ## Bug reports
 
-Please use the [Github issue tracking system for any bugs](https://github.com/grburgess/popsynth/issues), for questions, and or feature requests.
+Please use the [Github issue tracking system for any
+bugs](https://github.com/grburgess/popsynth/issues), for questions,
+and or feature requests.
 
 ## Code and more distributions
 
-While it is easy to create custom distributions in your local setup, if you would like to add them to popsynth directly, go ahead. Please include tests to ensure that your contributions are compatible with the code and can be maintained in the long term. 
+While it is easy to create custom distributions in your local setup,
+if you would like to add them to popsynth directly, go ahead. Please
+include tests to ensure that your contributions are compatible with
+the code and can be maintained in the long term.
 
 ## Documentation
 
-Additions or examples, tutorials, or better explanations are always welcome. To ensure that the documentation builds with the current version of the software, I am using (jupytext)[https://jupytext.readthedocs.io/en/latest/] to write the documentation in Markdown. These are automatically converted to and executed as jupyter notebooks when changes are pushed to Github.
+Additions or examples, tutorials, or better explanations are always
+welcome. To ensure that the documentation builds with the current
+version of the software, I am using
+[jupytext](https://jupytext.readthedocs.io/en/latest/) to write the
+documentation in Markdown. These are automatically converted to and
+executed as jupyter notebooks when changes are pushed to Github.
 
 
