+++ MANIFEST.in
@@ -1,2 +1,3 @@
 include versioneer.py
 include popsynth/_version.py
+graft popsynth/data
