### 2021-06-22 #3257 @grburgess
@HeloiseS and @warrickball I have worked a bit on the documentation to address you concerns. There were no issues opened so I will try to make sure and address every thing here:

I've added more in-depth explanation of the core of the procedure:

https://popsynth.readthedocs.io/en/latest/notebooks/distributions.html

This discusses the mathematics behind things. I could also add this to the paper, but from what I've seen in other JOSS papers, this might be a little too much... so I leave this up to the reviewers and editors.

I've added a realistic example for GRBS:

https://popsynth.readthedocs.io/en/latest/notebooks/examples.html#Short-GRBS

and there will be a blazar example by the end of the week.

Indeed, there are users that are developing some very different use cases for popsynth would be impossible to fully document as the abstraction of the framework opens up a lot of different possibilities. Some of these are currently being developed for publications in process, but I've asked those users to link in more examples when their publications are accepted.

The YAML functionality is no longer experimental. I've tried to add the test case data (pop.yml) so that it for sure is in the release, but I do not use conda and can't test that.

Luminosity functions are probability distributions. In popsynth they can be direct (e.g. sampling from a power law) or the luminosity function can be an emergent property of auxiliary samplers. This is documented here: https://popsynth.readthedocs.io/en/latest/notebooks/distributions.html#Creating-a-simple-population-synth and here: https://popsynth.readthedocs.io/en/latest/notebooks/aux.html#Derived-Luminosity-sampler

I have now documented that the 3D plots and graphs are available when their required packages are installed and configured. Unfortunately, these softwares are not the cleanest things to install. Nevertheless, these are only eye candy. As the docs are built live on each commit, their functionality is tested via continuous integration.

The API was failing because the numpy version with type hinting was not installed by default and I have now fixed this. I have tried to update all the doc strings, but there are still places where they are missing.  I will fill these out with time, but the main functionality has both doc strings and type hinting. Also, for some reason the inits are now not spitting out their doctoring with sphinx. I will try to clean this up as well.

> But why? I need a solid reason to add complexity.
> Also I don't understand the justification for the currently implemented Auxiliary Samplers: I do understand the ViewingAngle, >but what's the science behind having a log-normal or a log10 Auxiliary Sampler?

These features are used so that one can simulate very complex. For example, in one project relying on popsynth, people are simulating blazars from a luminosity and spatial distribution, as well as sampling their continuum and line spectra. All of these are random variables. Some quantities are sampled from log normal distributions. Rather than rewrite these every time, I have included them in the package.

> Also I am concerned that I don't see any validation: How do I know these give me realistic simulations? I liked your graphs showing the "selected" and "non-selected" objects - I'd like to see how that compares to other theoretical predictions. Say the feed the rate of supernovae from stellar evolution codes into popsynth and see if you roughly retrieve the number of observed supernovae (given a volume).

As the framework is generic, one would have to specify a theoretical problem to verify. As far as simple verification of the sampling, this is simply using Monte Carlo to integrate the equation now in the documentation. If you can provide an example, I can try and simulate that problem.

> Here are a couple of examples:
What is a SFR-like redshift? What is a Schecter luminosity function? What populations are we synthesysing? Stars? Transients? Observations? All three?

The SFR like is now explained in the documentation. Here is the some info on a schechter-luminosity function: https://ui.adsabs.harvard.edu/abs/1976ApJ...203..297S/abstract

The included distributions are just things that others and myself commonly have used. However, the reason for providing the ability to build your own outside of the repo, i.e., locally and on the fly in a notebook, is so that one can easily extend/ add to the framework as needed. The objects that are simulated can be anything.

I am updating the refs in the paper. Let me know if  there are remaining issues.
