### 2021-06-23 90a1ab9
added some power law samplers
