+++ popsynth/aux_samplers/__init__.py
@@ -2,8 +2,10 @@ from .viewing_angle_sampler import ViewingAngleSampler
 from .lognormal_aux_sampler import LogNormalAuxSampler, Log10NormalAuxSampler
 from .normal_aux_sampler import NormalAuxSampler
 from .trunc_normal_aux_sampler import TruncatedNormalAuxSampler
+from .plaw_aux_sampler import ParetoAuxSampler, PowerLawAuxSampler, BrokenPowerLawAuxSampler
+
 
 __all__ = [
     "ViewingAngleSampler", "LogNormalAuxSampler", "Log10NormalAuxSampler",
-    "NormalAuxSampler", "TruncatedNormalAuxSampler"
+    "NormalAuxSampler", "TruncatedNormalAuxSampler", "ParetoAuxSampler", "PowerLawAuxSampler", "BrokenPowerLawAuxSampler"
 ]
