### 2021-06-23 #3257 @warrickball
Hi @grburgess,

First, thanks for the [core concept](https://popsynth.readthedocs.io/en/latest/notebooks/distributions.html) page: it helps! I'm still not clear about the distinction between objects that always shine versus transients. The [short GRB example](https://popsynth.readthedocs.io/en/latest/notebooks/examples.html#Short-GRBS) makes the transient case clearer but what if I wanted to survey something like stars of a particular spectral type? My guess is that this is related to the definition of "intensity" as being some sort of time average and the only distinction between "transient" and "non-transient" is that the non-transient intensity is on average much greater. But it isn't clear to me how one might set this up in `popsynth`.

For one specific thing, how do I specify the duration Δt of the survey?

Perhaps I'm mistakenly interpreting the populations of the transient events as _progenitors_ of those events, so that e.g. some GRB progenitors don't become GRB during the survey and we don't detect them. But perhaps the point is that `popsynth` doesn't care about the actual rate of events, or at least that's somehow incorporated in the luminosity distribution and we have to specify something that makes sense according to a (maybe probabilistic) model.

I personally don't find the "fire fly" concept in the [quick start](https://popsynth.readthedocs.io/en/latest/notebooks/quickstart.html) has helped me understand any better. Do they light up continuously? Do they repeat? I find the astrophysical examples somewhat more relatable.

There are still gaps in the API which I think would help a lot in terms of making the distributions make sense. Some are simple fixes (e.g. [SFRDistribution](https://popsynth.readthedocs.io/en/latest/api/popsynth.distributions.html#popsynth.distributions.SFRDistribution) could link to the relevant article) but many could simply be fleshed out more with the relevant formulae.

----

To create a concrete example that I hoped would cement some of my misunderstandings, I thought about making a flux-limited sample of stars with luminosities L=M³, where M is the mass drawn from some supposed initial mass function (I ended up using a lognormal). The aim is to demonstrate that the surveyed sample will be biased towards higher masses than the underlying sample. I spent about an hour on this but found myself uncertain how to proceed. I think I need to use a [custom derived luminosity sampler](https://popsynth.readthedocs.io/en/latest/notebooks/aux.html#Derived-Luminosity-sampler), in which the underlying distribution would be the mass distribution, so I got as far as this script, largely copied from the docs:
```Python
#!/usr/bin/env python3

import numpy as np
import matplotlib.pyplot as pl
import popsynth

class LuminositySampler(popsynth.DerivedLumAuxSampler):
    _auxiliary_sampler_name = "LuminositySampler"
    mu = popsynth.auxiliary_sampler.AuxiliaryParameter(default=0)
    tau = popsynth.auxiliary_sampler.AuxiliaryParameter(default=1, vmin=0)

    def __init__(self, mu=0.0, tau=1.0, sigma=1):
        # this time set observed=True
        super(LuminositySampler, self).__init__("lumsample", uses_distance=False)

    def true_sampler(self, size):
        # draw a random number
        tmp = np.random.lognormal(self.mu, self.tau, size=size)
        self._true_values = tmp

    def compute_luminosity(self):
        # compute the luminosity
        return 10**(3*np.log10(self._true_values))

lum = LuminositySampler()

pop_gen = popsynth.populations.SphericalPopulation(1, r_max=10)
pop_gen.add_observed_quantity(lum)
flux_selector = popsynth.HardFluxSelection()
flux_selector.boundary = 1e-2
pop_gen.set_flux_selection(flux_selector)
pop = pop_gen.draw_survey(flux_sigma=0.1)

# pop.display_fluxes(obs_color='m', true_color='y', s=15)
# pl.show()

pl.hist(pop.luminosities_latent**(1/3), bins=np.arange(0.2,20.0,0.2), label='all');
pl.hist(pop.luminosities_latent[pop.selection]**(1/3), bins=np.arange(0.2,20.0,0.2), label='selected');
pl.xlabel('M')
pl.legend()
pl.show()
```
The fact that I've managed to get what looks like the right result (a flux-limited survey is biased towards more massive stars) means the docs are in a better state than my first impression suggested! But I don't pretend to understand the values (e.g. the flux boundary value) nor the relationship between the samplers, since I've simply copied the docs. I'm also not sure if there's a way to get to the underlying mass distribution instead of working it out from the transformation L=M³.
![image](https://user-images.githubusercontent.com/20858744/123099351-0440c280-d42a-11eb-8593-76a8f4f3bcc2.png)

I still think the API docs need the most work. I was mostly exploring what attributes and functions are available by hitting tab for autocomplete options in IPython. I was also querying docstrings using the `?` magic, which was often unhelpful. It would also be helpful to implement more `__repr__` and/or `__str__` functions, which would make it easier to see what attributes might not have been set, or quickly find out what default values they've taken. I've only used very simple distributions: the more complicated ones are more difficult because it's hard to see exactly what they do.

--

To finish with some concrete actions:
1. It still isn't clear how we relate the luminosities of transient versus non-transient things, and how the transient rate ("intensity"? "luminosity"?) incorporates the survey duration. The docs still need to explain this better upfront. It seems that to you there is no difference but to me there still is. (Consider: what about things that are intermittent and not strictly regular, like dwarf novae?)
2. I think the examples would benefit from a second case, aside from the GRB example, of something that isn't transient.
3. There are some heading issues where different headings from the same page are being listed in the contents sidebar. Not a big issue but a bit confusing for navigation.

I'll re-read the paper soon.
