### 2021-06-23 #3257 @grburgess
> Hi @grburgess,
>
> First, thanks for the [core concept](https://popsynth.readthedocs.io/en/latest/notebooks/distributions.html) page: it helps! I'm still not clear about the distinction between objects that always shine versus transients. The [short GRB example](https://popsynth.readthedocs.io/en/latest/notebooks/examples.html#Short-GRBS) makes the transient case clearer but what if I wanted to survey something like stars of a particular spectral type? My guess is that this is related to the definition of "intensity" as being some sort of time average and the only distinction between "transient" and "non-transient" is that the non-transient intensity is on average much greater. But it isn't clear to me how one might set this up in `popsynth`.

For the included cosmological distributions (though this could be modified externally), there is a constructor argument "as_rate" I haven't documented this well in the RTD page. But the difference here is on the dV/dz Jacobean that you need to account for the change in time on the cosmological interval when dealing with transients. If you aren't using cosmological distributions, then it depends on how you are define your volume elements.

https://github.com/grburgess/popsynth/blob/master/popsynth/distributions/cosmological_distribution.py#L11

https://github.com/grburgess/popsynth/blob/master/popsynth/distributions/cosmological_distribution.py#L50

This is  a little bit involved and does depend on what you are modeling.

Also, the definition of intensity is just being used from the more general form of Poisson where we are talking about counting in an N-dim space: https://en.wikipedia.org/wiki/Poisson_point_process

You can imagine that each object's process is a Poisson point process in its space. I just don't want to go too much into that here as it can be found in the references.

>
> For one specific thing, how do I specify the duration Δt of the survey?

It factors into the normalization of your spatial distribution. One would have to work this out for whatever case you are handling.

>
> Perhaps I'm mistakenly interpreting the populations of the transient events as _progenitors_ of those events, so that e.g. some GRB progenitors don't become GRB during the survey and we don't detect them. But perhaps the point is that `popsynth` doesn't care about the actual rate of events, or at least that's somehow incorporated in the luminosity distribution and we have to specify something that makes sense according to a (maybe probabilistic) model.

Yeah, there is no distinction... it really depends on how you define your volume elements.

>
> I personally don't find the "fire fly" concept in the [quick start](https://popsynth.readthedocs.io/en/latest/notebooks/quickstart.html) has helped me understand any better. Do they light up continuously? Do they repeat? I find the astrophysical examples somewhat more relatable.

I can remove this.

>
> There are still gaps in the API which I think would help a lot in terms of making the distributions make sense. Some are simple fixes (e.g. [SFRDistribution](https://popsynth.readthedocs.io/en/latest/api/popsynth.distributions.html#popsynth.distributions.SFRDistribution) could link to the relevant article) but many could simply be fleshed out more with the relevant formulae.

It is in the doc string but sphinx- auto doc is not working.

https://github.com/grburgess/popsynth/blob/master/popsynth/distributions/cosmological_distribution.py#L69

I think the better idea is to rename this to ColeSFR.. but this would break a lot of people's code at the moment. I think I will do that on a version bump.

>
> To create a concrete example that I hoped would cement some of my misunderstandings, I thought about making a flux-limited sample of stars with luminosities L=M³, where M is the mass drawn from some supposed initial mass function (I ended up using a lognormal). The aim is to demonstrate that the surveyed sample will be biased towards higher masses than the underlying sample. I spent about an hour on this but found myself uncertain how to proceed. I think I need to use a [custom derived luminosity sampler](https://popsynth.readthedocs.io/en/latest/notebooks/aux.html#Derived-Luminosity-sampler), in which the underlying distribution would be the mass distribution, so I got as far as this script, largely copied from the docs:
>
> ```python
> #!/usr/bin/env python3
>
> import numpy as np
> import matplotlib.pyplot as pl
> import popsynth
>
> class LuminositySampler(popsynth.DerivedLumAuxSampler):
>     _auxiliary_sampler_name = "LuminositySampler"
>     mu = popsynth.auxiliary_sampler.AuxiliaryParameter(default=0)
>     tau = popsynth.auxiliary_sampler.AuxiliaryParameter(default=1, vmin=0)
>
>     def __init__(self, mu=0.0, tau=1.0, sigma=1):
>         # this time set observed=True
>         super(LuminositySampler, self).__init__("lumsample", uses_distance=False)
>
>     def true_sampler(self, size):
>         # draw a random number
>         tmp = np.random.lognormal(self.mu, self.tau, size=size)
>         self._true_values = tmp
>
>     def compute_luminosity(self):
>         # compute the luminosity
>         return 10**(3*np.log10(self._true_values))
>
> lum = LuminositySampler()
>
> pop_gen = popsynth.populations.SphericalPopulation(1, r_max=10)
> pop_gen.add_observed_quantity(lum)
> flux_selector = popsynth.HardFluxSelection()
> flux_selector.boundary = 1e-2
> pop_gen.set_flux_selection(flux_selector)
> pop = pop_gen.draw_survey(flux_sigma=0.1)
>
> # pop.display_fluxes(obs_color='m', true_color='y', s=15)
> # pl.show()
>
> pl.hist(pop.luminosities_latent**(1/3), bins=np.arange(0.2,20.0,0.2), label='all');
> pl.hist(pop.luminosities_latent[pop.selection]**(1/3), bins=np.arange(0.2,20.0,0.2), label='selected');
> pl.xlabel('M')
> pl.legend()
> pl.show()
> ```
>
> The fact that I've managed to get what looks like the right result (a flux-limited survey is biased towards more massive stars) means the docs are in a better state than my first impression suggested! But I don't pretend to understand the values (e.g. the flux boundary value) nor the relationship between the samplers, since I've simply copied the docs. I'm also not sure if there's a way to get to the underlying mass distribution instead of working it out from the transformation L=M³.
> ![image](https://user-images.githubusercontent.com/20858744/123099351-0440c280-d42a-11eb-8593-76a8f4f3bcc2.png)

This is very cool! Perhaps you would like to add this to the examples? More examples of different things could help people think of how to use this. The boundary is for the value of flux in erg/s/cm2. I need to doc this.  You could also do this:

```python

mport numpy as np
import matplotlib.pyplot as pl
import popsynth

# create a sampler for mass
mass = popsynth.LogNormalAuxSampler(name="mass", observed = False)

class LuminositySampler(popsynth.DerivedLumAuxSampler):
    _auxiliary_sampler_name = "LuminositySampler"

    def __init__(self, mu=0.0, tau=1.0, sigma=1):
        # this time set observed=True
        super(LuminositySampler, self).__init__("lumsample", uses_distance=False)

    def true_sampler(self, size):

        mass = self._secondary_samplers["mass"].true_values

        self._true_values = mass

    def compute_luminosity(self):
        # compute the luminosity
        return 10**(3*np.log10(self._true_values))

lum = LuminositySampler()

lum.set_secondary_sampler(mass)

pop_gen = popsynth.populations.SphericalPopulation(1, r_max=10)
pop_gen.add_observed_quantity(lum)
flux_selector = popsynth.HardFluxSelection()
flux_selector.boundary = 1e-2
pop_gen.set_flux_selection(flux_selector)
pop = pop_gen.draw_survey(flux_sigma=0.1)

# pop.display_fluxes(obs_color='m', true_color='y', s=15)
# pl.show()

fig, ax = plt.subplots()

ax.hist(pop.mass, bins=np.arange(0.2,20.0,0.2), label='all');
ax.hist(pop.mass[pop.selection], bins=np.arange(0.2,20.0,0.2), label='selected');
ax.set_xlabel('M')
ax.legend()

pop.display_fluxes();

fig, ax = plt.subplots()

ax.hist(pop.luminosities_latent, bins=np.arange(0.2,20.0,0.2), label='all');
ax.hist(pop.luminosities_latent[pop.selection], bins=np.arange(0.2,20.0,0.2), label='selected');
ax.set_xlabel('L')
ax.legend()

```
![image](https://user-images.githubusercontent.com/6764007/123107275-acaa5300-d439-11eb-9b9d-3c3525d2ce10.png)

![image](https://user-images.githubusercontent.com/6764007/123107290-b0d67080-d439-11eb-914f-454a6b7db022.png)

![image](https://user-images.githubusercontent.com/6764007/123107301-b59b2480-d439-11eb-9e76-d86972f7e536.png)

>
> I still think the API docs need the most work. I was mostly exploring what attributes and functions are available by hitting tab for autocomplete options in IPython. I was also querying docstrings using the `?` magic, which was often unhelpful. It would also be helpful to implement more `__repr__` and/or `__str__` functions, which would make it easier to see what attributes might not have been set, or quickly find out what default values they've taken. I've only used very simple distributions: the more complicated ones are more difficult because it's hard to see exactly what they do.

The `?` should work on most of the constructors. But as I've said... autodoc is being a pain. I will work to fill out the docs strings more deeply in the code. all the distributions should have a display method which explains the math. But I've gotten behind on keeping up with this.

>
> --
>
> To finish with some concrete actions:
>
> 1. It still isn't clear how we relate the luminosities of transient versus non-transient things, and how the transient rate ("intensity"? "luminosity"?) incorporates the survey duration. The docs still need to explain this better upfront. It seems that to you there is no difference but to me there still is. (Consider: what about things that are intermittent and not strictly regular, like dwarf novae?)

I will doc the as_rate function better. But in general, it would be upon the user to write a correct spatial distribution. E.g., there are people using this to model flaring blazars, but these flare for much longer than GRBs. We have also used this to simulate static sources.

> 2. I think the examples would benefit from a second case, aside from the GRB example, of something that isn't transient.

I will add the blazar example soon. But again, your example is very nice.

> 3. There are some heading issues where different headings from the same page are being listed in the contents sidebar. Not a big issue but a bit confusing for navigation.

Ah... I need to separate this out.

>
> I'll re-read the paper soon.
