### 2021-06-23 #3257 @warrickball
Okay, I think I've (very slowly!) come around to how the transient versus non-transient thing works. So I think the missing piece of the documentation is a description (or link to a description of) how the relevant normalization constant should be calculated and which parameter it corresponds to in `popsynth`. At the moment I don't think this is mentioned at all in the docs.

You're more than welcome to use my example however you like. It'd benefit from being fleshed out with some vaguely reasonable numbers for the fluxes and distances and whatever other parameters. An interesting version—but not one I expect for this review!—might be to demonstrate how the (main sequence) stars we see by eye in the night sky are a flux-limited sample and therefore biased to higher masses. That could even be compared to a list of actual main sequence stars with V<4 or something (to keep the number small). I don't know how to construct that query but it's probably not too hard with astroquery.

--

Finally, regarding the formulae for the various functions, it's all well using the `display` method (though I don't see it implemented everywhere) but don't presume everyone uses the same workflow and tools as you. I don't routinely use notebooks and in my IPython console the maths just displays as `<IPython.core.display.Math object>`. I'd recommend just writing the formulae in the docstrings, much like e.g. [SciPy](https://docs.scipy.org/doc/scipy/reference/reference/generated/scipy.stats.pareto.html). It's not pretty in a console either but at least I can use my mental LaTeX parser, which isn't too bad after all these years!
