+++ popsynth/aux_samplers/__init__.py
@@ -4,8 +4,8 @@ from .normal_aux_sampler import NormalAuxSampler
 from .trunc_normal_aux_sampler import TruncatedNormalAuxSampler
 from .plaw_aux_sampler import ParetoAuxSampler, PowerLawAuxSampler, BrokenPowerLawAuxSampler
 
-
 __all__ = [
     "ViewingAngleSampler", "LogNormalAuxSampler", "Log10NormalAuxSampler",
-    "NormalAuxSampler", "TruncatedNormalAuxSampler", "ParetoAuxSampler", "PowerLawAuxSampler", "BrokenPowerLawAuxSampler"
+    "NormalAuxSampler", "TruncatedNormalAuxSampler", "ParetoAuxSampler",
+    "PowerLawAuxSampler", "BrokenPowerLawAuxSampler"
 ]
