+++ popsynth/aux_samplers/plaw_aux_sampler.py
@@ -5,6 +5,7 @@ import numba as nb
 from popsynth.auxiliary_sampler import AuxiliarySampler, AuxiliaryParameter
 from popsynth.distributions.bpl_distribution import bpl
 
+
 class ParetoAuxSampler(AuxiliarySampler):
     _auxiliary_sampler_name = "ParetoAuxSampler"
 
@@ -23,12 +24,13 @@ class ParetoAuxSampler(AuxiliarySampler):
         :returns: 
 
         """
-    
+
         super(ParetoAuxSampler, self).__init__(name=name, observed=observed)
 
     def true_sampler(self, size: int):
 
-         self._true_values = (np.random.pareto(self.alpha, size) + 1) * self.xmin
+        self._true_values = (np.random.pareto(self.alpha, size) +
+                             1) * self.xmin
 
     def observation_sampler(self, size: int):
 
@@ -41,6 +43,7 @@ class ParetoAuxSampler(AuxiliarySampler):
         else:
 
             self._obs_values = self._true_values
+
 
 class PowerLawAuxSampler(AuxiliarySampler):
     _auxiliary_sampler_name = "PowerLawAuxSampler"
@@ -61,12 +64,13 @@ class PowerLawAuxSampler(AuxiliarySampler):
         :returns: 
 
         """
-    
+
         super(PowerLawAuxSampler, self).__init__(name=name, observed=observed)
 
     def true_sampler(self, size: int):
 
-        self._true_values = _sample_power_law(self.xmin, self.xmax, self.alpha, size)
+        self._true_values = _sample_power_law(self.xmin, self.xmax, self.alpha,
+                                              size)
 
     def observation_sampler(self, size: int):
 
@@ -90,7 +94,6 @@ class BrokenPowerLawAuxSampler(AuxiliarySampler):
     beta = AuxiliaryParameter()
     xmax = AuxiliaryParameter(vmin=0)
 
-
     def __init__(self, name: str, observed: bool = True):
         """
         A power law distribution sampler
@@ -102,14 +105,16 @@ class BrokenPowerLawAuxSampler(AuxiliarySampler):
         :returns: 
 
         """
-    
-        super(BrokenPowerLawAuxSampler, self).__init__(name=name, observed=observed)
+
+        super(BrokenPowerLawAuxSampler, self).__init__(name=name,
+                                                       observed=observed)
 
     def true_sampler(self, size: int):
 
         u = np.atleast_1d(np.random.uniform(size=size))
-        
-        self._true_values = bpl(u, self.xmin, self.xbreak,self.xmax, self.alpha,self.beta)
+
+        self._true_values = bpl(u, self.xmin, self.xbreak, self.xmax,
+                                self.alpha, self.beta)
 
     def observation_sampler(self, size: int):
 
@@ -124,10 +129,6 @@ class BrokenPowerLawAuxSampler(AuxiliarySampler):
             self._obs_values = self._true_values
 
 
-
-    
-            
-
 @nb.njit(fastmath=True)
 def _sample_power_law(xmin, xmax, alpha, size):
 
@@ -137,16 +138,11 @@ def _sample_power_law(xmin, xmax, alpha, size):
 
         u = np.random.uniform(0, 1)
         x = np.power(
-            (np.power(xmax, alpha + 1) - np.power(xmin, alpha + 1)) * u
-            + np.power(xmin, alpha + 1),
+            (np.power(xmax, alpha + 1) - np.power(xmin, alpha + 1)) * u +
+            np.power(xmin, alpha + 1),
             1.0 / (alpha + 1.0),
-            )
+        )
 
         out[i] = x
 
-    
     return out
-
-
-
-    
