+++ popsynth/aux_samplers/__init__.py
@@ -1,11 +1,23 @@
+from .delta_aux_sampler import DeltaAuxSampler
 from .viewing_angle_sampler import ViewingAngleSampler
 from .lognormal_aux_sampler import LogNormalAuxSampler, Log10NormalAuxSampler
 from .normal_aux_sampler import NormalAuxSampler
 from .trunc_normal_aux_sampler import TruncatedNormalAuxSampler
-from .plaw_aux_sampler import ParetoAuxSampler, PowerLawAuxSampler, BrokenPowerLawAuxSampler
+from .plaw_aux_sampler import (
+    ParetoAuxSampler,
+    PowerLawAuxSampler,
+    BrokenPowerLawAuxSampler,
+)
 
 
 __all__ = [
-    "ViewingAngleSampler", "LogNormalAuxSampler", "Log10NormalAuxSampler",
-    "NormalAuxSampler", "TruncatedNormalAuxSampler", "ParetoAuxSampler", "PowerLawAuxSampler", "BrokenPowerLawAuxSampler"
+    "DeltaAuxSampler",
+    "ViewingAngleSampler",
+    "LogNormalAuxSampler",
+    "Log10NormalAuxSampler",
+    "NormalAuxSampler",
+    "TruncatedNormalAuxSampler",
+    "ParetoAuxSampler",
+    "PowerLawAuxSampler",
+    "BrokenPowerLawAuxSampler",
 ]
