### 2021-06-25 af3ff22
Docstrings for aux_samplers
