+++ popsynth/aux_samplers/viewing_angle_sampler.py
@@ -1,7 +1,6 @@
 import numpy as np
 
-from popsynth.auxiliary_sampler import (AuxiliaryParameter,
-                                        NonObservedAuxSampler)
+from popsynth.auxiliary_sampler import AuxiliaryParameter, NonObservedAuxSampler
 
 
 class ViewingAngleSampler(NonObservedAuxSampler):
@@ -11,16 +10,18 @@ class ViewingAngleSampler(NonObservedAuxSampler):
 
     def __init__(self):
         """
-        A viewing angle sampler that samples from 0, max_angle.
-        It assumes that this is NOT an observed property
+        A viewing angle sampler that samples
+        from 0 to ``max_angle``. Unlike other samplers,
+        it assumes that this is NOT an observed property
 
-        :param max_angle: the maximum angle to which to sample in DEGS
-        :returns: None
-        :rtype: None
-
+        :param max_angle: The maximum angle to which to
+        sample in degrees
+        :type max_angle: :class:`AuxiliaryParameter`
         """
 
-        super(ViewingAngleSampler, self).__init__(name="va", )
+        super(ViewingAngleSampler, self).__init__(
+            name="va",
+        )
 
     def true_sampler(self, size: int) -> None:
         """
@@ -32,9 +33,8 @@ class ViewingAngleSampler(NonObservedAuxSampler):
 
         """
 
-        theta_inverse = np.random.uniform(0.0,
-                                          1 -
-                                          np.cos(np.deg2rad(self.max_angle)),
-                                          size=size)
+        theta_inverse = np.random.uniform(
+            0.0, 1 - np.cos(np.deg2rad(self.max_angle)), size=size
+        )
 
         self._true_values = np.arccos(1.0 - theta_inverse)
