+++ popsynth/aux_samplers/plaw_aux_sampler.py
@@ -5,6 +5,7 @@ import numba as nb
 from popsynth.auxiliary_sampler import AuxiliarySampler, AuxiliaryParameter
 from popsynth.distributions.bpl_distribution import bpl
 
+
 class ParetoAuxSampler(AuxiliarySampler):
     _auxiliary_sampler_name = "ParetoAuxSampler"
 
@@ -14,33 +15,42 @@ class ParetoAuxSampler(AuxiliarySampler):
 
     def __init__(self, name: str, observed: bool = True):
         """
-        A pareto distribution sampler
+        A pareto distribution sampler,
+        where property ~ 1 / x^(``alpha`` + 1).
 
-        :param name: 
+        :param name: Name of the property
         :type name: str
-        :param observed: 
+        :param observed: `True` if the property is observed,
+        `False` if it is latent. Defaults to `True`
         :type observed: bool
-        :returns: 
-
+        :param xmin: Minimum value of the pareto
+        :type xmin: :class:`AuxiliaryParameter`
+        :param alpha: Index of the pareto
+        :type alpha: :class:`AuxiliaryParameter`
+        :param sigma: Standard deviation of normal distribution
+        from which observed values are sampled, if ``observed``
+        is `True`
+        :type sigma: :class:`AuxiliaryParameter`
         """
-    
+
         super(ParetoAuxSampler, self).__init__(name=name, observed=observed)
 
     def true_sampler(self, size: int):
 
-         self._true_values = (np.random.pareto(self.alpha, size) + 1) * self.xmin
+        self._true_values = (np.random.pareto(self.alpha, size) + 1) * self.xmin
 
     def observation_sampler(self, size: int):
 
         if self._is_observed:
 
-            self._obs_values = stats.norm.rvs(loc=self._true_values,
-                                              scale=self.sigma,
-                                              size=size)
+            self._obs_values = stats.norm.rvs(
+                loc=self._true_values, scale=self.sigma, size=size
+            )
 
         else:
 
             self._obs_values = self._true_values
+
 
 class PowerLawAuxSampler(AuxiliarySampler):
     _auxiliary_sampler_name = "PowerLawAuxSampler"
@@ -52,16 +62,24 @@ class PowerLawAuxSampler(AuxiliarySampler):
 
     def __init__(self, name: str, observed: bool = True):
         """
-        A power law distribution sampler
+        A bounded power law distribution sampler,
+        where property ~ x^``alpha``.
 
-        :param name: 
+        :param name: Name of the property
         :type name: str
-        :param observed: 
+        :param observed: `True` if the property is observed,
+        `False` if it is latent. Defaults to `True`
         :type observed: bool
-        :returns: 
-
+        :param xmin: Minimum value of the power law
+        :type xmin: :class:`AuxiliaryParameter`
+        :param xmax: Maximum value of the power law
+        :type xmax: :class:``AuxiliaryParameter
+        :param sigma: Standard deviation of normal distribution
+        from which observed values are sampled, if ``observed``
+        is `True`
+        :type sigma: :class:`AuxiliaryParameter`
         """
-    
+
         super(PowerLawAuxSampler, self).__init__(name=name, observed=observed)
 
     def true_sampler(self, size: int):
@@ -72,9 +90,9 @@ class PowerLawAuxSampler(AuxiliarySampler):
 
         if self._is_observed:
 
-            self._obs_values = stats.norm.rvs(loc=self._true_values,
-                                              scale=self.sigma,
-                                              size=size)
+            self._obs_values = stats.norm.rvs(
+                loc=self._true_values, scale=self.sigma, size=size
+            )
 
         else:
 
[…]
