+++ popsynth/aux_samplers/sky_sampler.py
@@ -1,79 +1,84 @@
-# import numpy as np
+import numpy as np
 
-# from popsynth.auxiliary_sampler import NonObservedAuxSampler
-# from popsynth.utils.spherical_geometry import sample_theta_phi
+from popsynth.auxiliary_sampler import NonObservedAuxSampler
+from popsynth.utils.spherical_geometry import sample_theta_phi
 
-# class SkySampler(object):
-#     _auxiliary_sampler_name = "SkySampler"
 
-#     def __init__(self, ra_sampler=None, dec_sampler=None):
-#         """
-#         A sky sampler that samples theta phi positions
+class SkySampler(object):
+    _auxiliary_sampler_name = "SkySampler"
 
-#         :returns:
-#         :rtype:
+    def __init__(
+        self,
+        ra_sampler: NonObservedAuxSampler = None,
+        dec_sampler: NonObservedAuxSampler = None,
+    ):
+        """
+        A sky sampler that samples angular positions
+        in ra and dec. If no samplers are provided, then
+        loads default samplers that sample uniformly on
+        the unit sphere. RA and dec are in radians.
 
-#         """
+        :param ra_sampler: Right ascension (RA) sampler
+        :type ra_sampler: :class:`NonObservedAuxSampler`
+        :param dec_sampler: Declination (Dec) sampler
+        :type dec_sampler: :class:`NonObservedAuxSampler`
+        """
 
-#         self._ra_sampler = ra_sampler
-#         self._dec_sampler = dec_sampler
+        self._ra_sampler = ra_sampler
+        self._dec_sampler = dec_sampler
 
-#         self._setup_sky()
+        self._setup_sky()
 
-#     def _setup_sky(self):
+    def _setup_sky(self):
 
-#         if self._ra_sampler is None:
-#             self._ra_sampler = RASampler()
+        if self._ra_sampler is None:
+            self._ra_sampler = RASampler()
 
-#         if self._dec_sampler is None:
-#             self._dec_sampler = DecSampler()
+        if self._dec_sampler is None:
+            self._dec_sampler = DecSampler()
 
-#     @property
-#     def ra_sampler(self):
+    @property
+    def ra_sampler(self):
 
-#         return self._ra_sampler
+        return self._ra_sampler
 
-#     @property
-#     def dec_sampler(self):
+    @property
+    def dec_sampler(self):
 
-#         return self._dec_sampler
+        return self._dec_sampler
 
-# class RASampler(NonObservedAuxSampler):
-#     _auxiliary_sampler_name = "RASampler"
 
-#     def __init__(self):
-#         """
+class RASampler(NonObservedAuxSampler):
+    _auxiliary_sampler_name = "RASampler"
 
-#         Samples the RA of sky uniformly
+    def __init__(self):
+        """
+        Samples the right ascension (RA)
+        uniformly on the unit sphere.
 
-#         :param true_values:
-#         :returns:
-#         :rtype:
+        RA is in radians.
+        """
 
-#         """
+        super(RASampler, self).__init__(name="ra")
 
-#         super(RASampler, self).__init__(name="ra")
+    def true_sampler(self, size):
 
-#     def true_sampler(self, size):
+        self._true_values = np.random.uniform(0, 2 * np.pi, size=size)
 
-#         self._true_values = np.random.uniform(0, 2 * np.pi, size=size)
 
-# class DecSampler(NonObservedAuxSampler):
-#     _auxiliary_sampler_name = "DecSampler"
+class DecSampler(NonObservedAuxSampler):
+    _auxiliary_sampler_name = "DecSampler"
 
-#     def __init__(self):
-#         """
-#         sampler the dec of the sky uniformly
+    def __init__(self):
+        """
[…]
