### 2021-06-26 fe9fa2e
docstrings and lf_form for bpl_distribution
