### 2021-06-26 8f8bec0
Docstrings for cosmological_distribution including units + tidy bpl
