+++ popsynth/distributions/cosmological_distribution.py
@@ -1,7 +1,6 @@
 import numpy as np
 import numba as nb
-import math
-
+from typing import Any
 from popsynth.distribution import SpatialDistribution, DistributionParameter
 from popsynth.utils.cosmology import cosmology
 
@@ -9,28 +8,30 @@ from popsynth.utils.cosmology import cosmology
 class CosmologicalDistribution(SpatialDistribution):
     _distribution_name = "CosmologicalDistribution"
 
-    def __init__(self,
-                 seed=1234,
-                 name="cosmo",
-                 form=None,
-                 truth={},
-                 is_rate=True):
-
+    def __init__(
+        self,
+        seed: int = 1234,
+        name: str = "cosmo",
+        form: str = None,
+        truth: dict[str, Any] = {},
+        is_rate: bool = True,
+    ):
         """
-        base cosmological spatial distribution
+        Base class for cosmological spatial distributions.
 
-        :param seed: 
-        :type seed: 
-        :param name: 
-        :type name: 
-        :param form: 
-        :type form: 
-        :param truth: 
-        :type truth: 
-        :param is_rate: 
-        :type is_rate: 
-        :returns: 
-
+        :param seed: Random seed
+        :type seed: int
+        :param name: Name of the distribution
+        :type name: str
+        :param form: Mathematical description of distribution
+        :type form: str
+        :param truth: True values of parameters
+        :type truth: dict[str, Any]
+        :param is_rate: `True` if modelling a population of transient events,
+        `False` if modelling a population of steady-state objects.
+        Affects the ``time_adjustment`` method used in cosmo calculations.
+        Default is `True`.
+        :type is_rate: bool
         """
         super(CosmologicalDistribution, self).__init__(
             seed=seed,
@@ -40,17 +41,48 @@ class CosmologicalDistribution(SpatialDistribution):
         self._is_rate = is_rate
 
     def differential_volume(self, z):
+        """
+        Differential comoving volume in Gpc^3 sr^-1.
+
+        dV/dzdOmega
+
+        :param z: Redshift
+        :returns: The differential comoving volume in
+        Gpc^-3 sr^-1.
+        """
 
         return cosmology.differential_comoving_volume(z)
 
     def transform(self, L, z):
+        """
+        Transformation from luminosity to energy flux.
 
-        return L / (4.0 * np.pi * cosmology.luminosity_distance(z)**2)
+        L / 4 pi dL^2
+
+        dL is in cm. Therefore for L in erg s^-1 returns
+        flux in erg cm^-2 s^-1.
+
+        :param L: Luminosity
+        :param z: Redshift
+        :returns: Flux
+        """
+
+        return L / (4.0 * np.pi * cosmology.luminosity_distance(z) ** 2)
 
     def time_adjustment(self, z):
+        """
+        Time adjustment factor to handle both
+        transient and steady-state populations.
+
+        :param z: Redshift
+        :returns: Appropriate factor depending on ``is_rate``
+        """
         if self._is_rate:
+
             return 1 + z
+
         else:
+
             return 1.0
 
 
@@ -63,22 +95,34 @@ class SFRDistribution(CosmologicalDistribution):
     decay = DistributionParameter()
     peak = DistributionParameter(vmin=0)
 
-    def __init__(self, seed=1234, name="sfr", is_rate=True):
-
+    def __init__(self, seed: int = 1234, name: str = "sfr", is_rate: bool = True):
         """
         A star-formation like distribution of the form
-        presented in Cole et al. 2001
+        presented in Cole et al. 2001.
 
-        :param seed: 
[…]
