### 2021-06-26 dcdc33c
Docstrings for spatial populations
