+++ popsynth/populations/spatial_populations.py
@@ -1,30 +1,41 @@
 from popsynth.distributions.cosmological_distribution import (  # MergerDistribution,
-    SFRDistribution, ZPowerCosmoDistribution)
+    SFRDistribution,
+    ZPowerCosmoDistribution,
+)
 from popsynth.distributions.spherical_distribution import (
-    ConstantSphericalDistribution, ZPowerSphericalDistribution)
-from popsynth.distributions.spiral_galaxy_distribution import \
-    SpiralGalaxyDistribution
+    ConstantSphericalDistribution,
+    ZPowerSphericalDistribution,
+)
+from popsynth.distribution import LuminosityDistribution
+from popsynth.distributions.spiral_galaxy_distribution import SpiralGalaxyDistribution
 from popsynth.population_synth import PopulationSynth
 
-# First create a bunch of spatial distributions
+"""
+Create a range of spatial populations that can
+be expanded with luminosity distributions.
+"""
 
 
 class SphericalPopulation(PopulationSynth):
-    def __init__(self,
-                 Lambda,
-                 r_max=5.0,
-                 seed=1234,
-                 luminosity_distribution=None):
+    def __init__(
+        self,
+        Lambda: float,
+        r_max: float = 5.0,
+        seed: int = 1234,
+        luminosity_distribution: LuminosityDistribution = None,
+    ):
         """
-        A generic spherical population
+        A generic spherical population based on
+        :class:`ConstantSphericalDistribution`.
 
-        :param Lambda:
-        :param r_max:
-        :param seed:
-        :param luminosity_distribution:
-        :returns:
-        :rtype:
-
+        :param Lambda: Density per unit volume
+        :type Lambda: float
+        :param r_max: Maximum distance
+        :type r_max: float
+        :param seed: Random seed
+        :type seed: int
+        :param luminosity_distribution: Luminosity distribution
+        :type luminosity_distribution: :class:`LuminosityDistribution`, optional
         """
 
         spatial_distribution = ConstantSphericalDistribution(seed=seed)
@@ -39,25 +50,29 @@ class SphericalPopulation(PopulationSynth):
 
 
 class ZPowerSphericalPopulation(PopulationSynth):
-    def __init__(self,
-                 Lambda,
-                 delta,
-                 r_max=5.0,
-                 seed=1234,
-                 luminosity_distribution=None):
+    def __init__(
+        self,
+        Lambda: float,
+        delta: float,
+        r_max: float = 5.0,
+        seed: int = 1234,
+        luminosity_distribution: LuminosityDistribution = None,
+    ):
         """
-        
-        A spherical population with a density that 
-        scales as (r+1)^delta
+        A spherical population with a density that
+        scales as (r+1)^delta.
+        Based on :class:`ZpowerSphericalDistribution`.
 
-        :param Lambda:
-        :param delta:
-        :param r_max:
-        :param seed:
-        :param luminosity_distribution:
-        :returns:
-        :rtype:
-
+        :param Lambda: Local density per unit volume
+        :type Lambda: float
+        :param delta: Index of spatial distribution
+        :type delta: float
+        :param r_max: Maximum distance
+        :type r_max: float
+        :param seed: Random seed
+        :type seed: int
+        :param luminosity_distribution: Luminosity distribution
+        :type luminosity_distribution: :class:`LuminosityDistribution`, optional
         """
 
         spatial_distribution = ZPowerSphericalDistribution(seed=seed)
@@ -76,27 +91,39 @@ class ZPowerSphericalPopulation(PopulationSynth):
 class ZPowerCosmoPopulation(PopulationSynth):
     def __init__(
         self,
-        Lambda,
-        delta,
-        r_max=5.0,
-        seed=1234,
-        luminosity_distribution=None,
-        is_rate=True,
+        Lambda: float,
[…]
