+++ popsynth/populations/bpl_population.py
@@ -8,28 +8,38 @@ from popsynth.distributions.bpl_distribution import BPLDistribution
 
 
 class BPLHomogeneousSphericalPopulation(SphericalPopulation):
-    def __init__(self,
-                 Lambda,
-                 Lmin,
-                 alpha,
-                 Lbreak,
-                 beta,
-                 Lmax,
-                 r_max=5,
-                 seed=1234):
-        """FIXME! briefly describe function
+    def __init__(
+        self,
+        Lambda: float,
+        Lmin: float,
+        alpha: float,
+        Lbreak: float,
+        beta: float,
+        Lmax: float,
+        r_max: float = 5,
+        seed: int = 1234,
+    ):
+        """
+        A population built on the :class:`ConstantSphericalDistribution`
+        spatial distribution and the :class:`BPLDistribution`
+        luminosity distribution.
 
-        :param Lambda:
-        :param Lmin:
-        :param alpha:
-        :param Lbreak:
-        :param beta:
-        :param Lmax:
-        :param r_max:
-        :param seed:
-        :returns:
-        :rtype:
-
+        :param Lambda: Density per unit volume
+        :type Lambda: float
+        :param Lmin: Minimum value of the luminosity
+        :type Lmin: float
+        :param alpha: Lower luminosity index
+        :type alpha: float
+        :param Lbreak: Break luminosity
+        :type Lbreak: float
+        :param beta: Upper luminosity index
+        :type beta: float
+        :param Lmax: Maximum value of the luminosity
+        :type Lmax: float
+        :param r_max: Maximum distance
+        :type r_max: float
+        :param seed: Random seed
+        :type seed: int
         """
         luminosity_distribution = BPLDistribution(seed=seed)
         luminosity_distribution.Lmin = Lmin
@@ -47,30 +57,41 @@ class BPLHomogeneousSphericalPopulation(SphericalPopulation):
 
 
 class BPLZPowerSphericalPopulation(ZPowerSphericalPopulation):
-    def __init__(self,
-                 Lambda,
-                 delta,
-                 Lmin,
-                 alpha,
-                 Lbreak,
-                 beta,
-                 Lmax,
-                 r_max=5,
-                 seed=1234):
-        """FIXME! briefly describe function
+    def __init__(
+        self,
+        Lambda: float,
+        delta: float,
+        Lmin: float,
+        alpha: float,
+        Lbreak: float,
+        beta: float,
+        Lmax: float,
+        r_max: float = 5,
+        seed: int = 1234,
+    ):
+        """
+        A population built on the :class:`ZPowerSphericalDistribution`
+        spatial distribution and the :class:`BPLDistribution`
+        luminosity distribution.
 
-        :param Lambda:
-        :param delta:
-        :param Lmin:
-        :param alpha:
-        :param Lbreak:
-        :param beta:
-        :param Lmax:
-        :param r_max:
-        :param seed:
-        :returns:
-        :rtype:
-
+        :param Lambda: Density per unit volume
+        :type Lambda: float
+        :param delta: Index of spatial distribution
+        :type delta: float
+        :param Lmin: Minimum value of the luminosity
+        :type Lmin: float
+        :param alpha: Lower luminosity index
+        :type alpha: float
+        :param Lbreak: Break luminosity
+        :type Lbreak: float
+        :param beta: Upper luminosity index
+        :type beta: float
+        :param Lmax: Maximum value of the luminosity
+        :type Lmax: float
+        :param r_max: Maximum distance
+        :type r_max: float
+        :param seed: Random seed
+        :type seed: int
         """
         luminosity_distribution = BPLDistribution(seed=seed)
[…]
