### 2021-06-26 6bff07f
Docstring for lognormal_populations
