+++ popsynth/populations/lognormal_population.py
@@ -10,17 +10,29 @@ from popsynth.distributions.log_normal_distribution import LogNormalDistribution
 
 
 class LogNormalHomogeneousSphericalPopulation(SphericalPopulation):
-    def __init__(self, Lambda, mu, tau, r_max=5, seed=1234):
-        """FIXME! briefly describe function
+    def __init__(
+        self,
+        Lambda: float,
+        mu: float,
+        tau: float,
+        r_max: float = 5,
+        seed: int = 1234,
+    ):
+        """
+        A population built on the :class:`ConstantSphericalDistribution`
+        spatial distribution and the :class:`LogNormalDistribution`
+        luminosity distribution.
 
-        :param Lambda:
-        :param mu:
-        :param tau:
-        :param r_max:
-        :param seed:
-        :returns:
-        :rtype:
-
+        :param Lambda: Density per unit volume
+        :type Lambda: float
+        :param mu: Mean of the luminosity distribution
+        :type mu: float
+        :param tau: Standard deviation of the luminosity distribution
+        :type tau: float
+        :param r_max: Maximum distance
+        :type r_max: float
+        :param seed: Random seed
+        :type seed: int
         """
 
         luminosity_distribution = LogNormalDistribution(seed=seed)
@@ -36,18 +48,32 @@ class LogNormalHomogeneousSphericalPopulation(SphericalPopulation):
 
 
 class LogNormalZPowerSphericalPopulation(ZPowerSphericalPopulation):
-    def __init__(self, Lambda, delta, mu, tau, r_max=5, seed=1234):
-        """FIXME! briefly describe function
+    def __init__(
+        self,
+        Lambda: float,
+        delta: float,
+        mu: float,
+        tau: float,
+        r_max: float = 5,
+        seed: int = 1234,
+    ):
+        """
+        A population built on the :class:`ZPowerSphericalDistribution`
+        spatial distribution and the :class:`LogNormalDistribution`
+        luminosity distribution.
 
-        :param Lambda:
-        :param delta:
-        :param mu:
-        :param tau:
-        :param r_max:
-        :param seed:
-        :returns:
-        :rtype:
-
+        :param Lambda: Density per unit volume
+        :type Lambda: float
+        :param delta: Index of the spatial distribution
+        :type delta: float
+        :param mu: Mean of the luminosity distribution
+        :type mu: float
+        :param tau: Standard deviation of the luminosity distribution
+        :type tau: float
+        :param r_max: Maximum distance
+        :type r_max: float
+        :param seed: Random seed
+        :type seed: int
         """
 
         luminosity_distribution = LogNormalDistribution(seed=seed)
@@ -66,25 +92,31 @@ class LogNormalZPowerSphericalPopulation(ZPowerSphericalPopulation):
 class LogNormalZPowerCosmoPopulation(ZPowerCosmoPopulation):
     def __init__(
         self,
-        Lambda,
-        delta,
-        mu,
-        tau,
-        r_max=5,
-        seed=1234,
-        is_rate=True,
+        Lambda: float,
+        delta: float,
+        mu: float,
+        tau: float,
+        r_max: float = 5,
+        seed: int = 1234,
+        is_rate: bool = True,
     ):
-        """FIXME! briefly describe function
+        """
+        A population built on the :class:`ZPowerCosmoDistribution`
+        spatial distribution and the :class:`LogNormalDistribution`
+        luminosity distribution.
 
-        :param Lambda:
-        :param delta:
-        :param mu:
-        :param tau:
-        :param r_max:
-        :param seed:
-        :returns:
-        :rtype:
-
+        :param Lambda: Density in units of Gpc^-3
+        :type Lambda: float
+        :param delta: Index of the spatial distribution
+        :type delta: float
[…]
