### 2021-06-26 36252a9
Docstrings for generic_selectors
