### 2021-06-26 d1556c8
Docstrings for all selection_probabilities
