+++ popsynth/selection_probability/spatial_selection.py
@@ -7,16 +7,26 @@ class SpatialSelection(SelectionProbabilty):
 
     _selection_name = "SpatialSelection"
 
-    def __init__(self, name) -> None:
+    def __init__(self, name: str) -> None:
+        """
+        A generic spatial selection.
+
+        :param name: Name of the selection
+        :type name: str
+        """
 
         super(SpatialSelection, self).__init__(name)
 
         self._spatial_distribution: SpatialDistribution = None
 
     def set_spatial_distribution(
-            self, spatial_distribtuion: SpatialDistribution) -> None:
+        self, spatial_distribtuion: SpatialDistribution
+    ) -> None:
         """
-        set the spatial distribution
+        Set the spatial distribution for the selection.
+
+        :param spatial_distribution: The spatial_distribution
+        :type spatial_distribution: :class:`SpatialDistribution`
         """
         assert isinstance(spatial_distribtuion, SpatialDistribution)
 
