### 2021-06-26 0527110
Docstrings for utils
