### 2021-06-26 94fab6d
auxiliary sampler docstrings
