+++ popsynth/auxiliary_sampler.py
@@ -21,15 +21,33 @@ class AuxiliaryParameter(Parameter):
     pass
 
 
-class AuxiliarySampler(object,
-                       metaclass=AutoRegister(auxiliary_parameter_registry,
-                                              base_type=ParameterMeta)):
-    def __init__(self,
-                 name: str,
-                 observed: bool = True,
-                 uses_distance: bool = False,
-                 uses_luminosity: bool = False,
-                 uses_sky_position: bool = False) -> None:
+class AuxiliarySampler(
+    object,
+    metaclass=AutoRegister(auxiliary_parameter_registry, base_type=ParameterMeta),
+):
+    def __init__(
+        self,
+        name: str,
+        observed: bool = True,
+        uses_distance: bool = False,
+        uses_luminosity: bool = False,
+        uses_sky_position: bool = False,
+    ) -> None:
+        """
+        Base class for auxiliary samplers.
+
+        :param name: Name of the sampler
+        :type name: str
+        :param observed: `True` if the property is observed,
+        `False` if it is latent. Defaults to `True`
+        :type observed: bool
+        :param uses_distance: `True` if sampler uses distance values
+        :type uses_distance: bool
+        :param uses_luminosity: `True` if sampler uses luminosities
+        :type uses_luminosity: bool
+        :param uses_sky_position: `True` if sampler uses sky positions
+        :type uses_sky_position: bool
+        """
 
         self._parameter_storage = {}  # type: Dict[str, float]
         self._name = name  # type: str
@@ -49,23 +67,21 @@ class AuxiliarySampler(object,
         self._uses_sky_position = uses_sky_position  # type: bool
 
     def set_luminosity(self, luminosity: ArrayLike) -> None:
-        """FIXME! briefly describe function
+        """
+        Set the luminosity values.
 
-        :param luminosity:
-        :returns:
-        :rtype:
-
+        :param luminosity: Luminosity
+        :type luminosity: ArrayLike
         """
 
         self._luminosity = luminosity  # type:ArrayLike
 
     def set_spatial_values(self, value: SpatialContainer) -> None:
-        """FIXME! briefly describe function
+        """
+        Set the spatial values.
 
-        :param distance:
-        :returns:
-        :rtype:
-
+        :param value: Spatial values
+        :type value: :class:`SpatialContainer`
         """
 
         self._distance = value.distance  # type:ArrayLike
@@ -85,14 +101,14 @@ class AuxiliarySampler(object,
 
     def _apply_selection(self) -> None:
         """
-        Default selection if none is specfied in child class
+        Default selection if none is specfied in child class.
         """
 
         self._selector.draw(len(self._obs_values))
 
     def set_secondary_sampler(self, sampler) -> None:
         """
-        Allows the setting of a secondary sampler from which to derive values
+        Allows the setting of a secondary sampler from which to derive values.
         """
 
         # make sure we set the sampler as a secondary
@@ -109,7 +125,8 @@ class AuxiliarySampler(object,
         """
         Draw the primary and secondary samplers. This is the main call.
 
-        :param size: the number of samples to draw
+        :param size: The number of samples to draw
+        :type size: int
         """
         # do not resample!
         if not self._is_sampled:
@@ -163,17 +180,17 @@ class AuxiliarySampler(object,
 
             else:
 
-                self._obs_values = (self._true_values)  # type: ArrayLike
+                self._obs_values = self._true_values  # type: ArrayLike
 
             self._selector.set_observed_value(self._obs_values)
 
             # check to make sure we sampled!
-            assert (self.true_values is not None
-                    and len(self.true_values) == size
[…]
