### 2021-06-26 a404cce
Docstrings for distribution
