+++ popsynth/population.py
@@ -1,9 +1,7 @@
-import importlib
-from typing import Any, Dict, List, Optional, Type
+from typing import Any, Dict, Optional
 
 import h5py
 import ipyvolume as ipv
-import ipywidgets as widgets
 import matplotlib.pyplot as plt
 import networkx as nx
 import numpy as np
@@ -15,71 +13,99 @@ from numpy.typing import ArrayLike
 
 from popsynth.utils.array_to_cmap import array_to_cmap
 from popsynth.utils.hdf5_utils import (
-    clean_graph_dict, fill_graph_dict,
+    clean_graph_dict,
+    fill_graph_dict,
     recursively_load_dict_contents_from_group,
-    recursively_save_dict_contents_to_group)
+    recursively_save_dict_contents_to_group,
+)
 from popsynth.utils.logging import setup_logger
 from popsynth.utils.spherical_geometry import xyz
 
 log = setup_logger(__name__)
 
 wine = "#8F2727"
-dark, dark_highlight, mid, mid_highlight, light, light_highlight = betagen(
-    wine)
+dark, dark_highlight, mid, mid_highlight, light, light_highlight = betagen(wine)
 
 
 class Population(object):
-    def __init__(self,
-                 luminosities: ArrayLike,
-                 distances: ArrayLike,
-                 known_distances: ArrayLike,
-                 known_distance_idx: ArrayLike,
-                 unknown_distance_idx: ArrayLike,
-                 fluxes: ArrayLike,
-                 flux_obs: ArrayLike,
-                 selection: ArrayLike,
-                 flux_sigma: float,
-                 r_max: float,
-                 n_model: int,
-                 lf_params: Dict[str, Any],
-                 spatial_params: Optional[Dict[str, Any]] = None,
-                 model_spaces: Optional[ArrayLike] = None,
-                 seed: int = 1234,
-                 name: Optional[str] = None,
-                 spatial_form: Optional[Dict[str, Any]] = None,
-                 lf_form: Optional[Dict[str, Any]] = None,
-                 auxiliary_quantities: Optional[Dict[str, Any]] = None,
-                 truth: Dict[str, float] = {},
-                 graph: Optional[Dict[str, Any]] = None,
-                 theta=None,
-                 phi=None,
-                 pop_synth: Optional[Dict[str, Any]] = None) -> None:
+    def __init__(
+        self,
+        luminosities: ArrayLike,
+        distances: ArrayLike,
+        known_distances: ArrayLike,
+        known_distance_idx: ArrayLike,
+        unknown_distance_idx: ArrayLike,
+        fluxes: ArrayLike,
+        flux_obs: ArrayLike,
+        selection: ArrayLike,
+        flux_sigma: float,
+        r_max: float,
+        n_model: int,
+        lf_params: Dict[str, Any],
+        spatial_params: Optional[Dict[str, Any]] = None,
+        model_spaces: Optional[ArrayLike] = None,
+        seed: int = 1234,
+        name: Optional[str] = None,
+        spatial_form: Optional[Dict[str, Any]] = None,
+        lf_form: Optional[Dict[str, Any]] = None,
+        auxiliary_quantities: Optional[Dict[str, Any]] = None,
+        truth: Dict[str, float] = {},
+        graph: Optional[Dict[str, Any]] = None,
+        theta=None,
+        phi=None,
+        pop_synth: Optional[Dict[str, Any]] = None,
+    ) -> None:
         """
         A population container for all the properties of the population.
 
-        :param luminosities: the luminosities
-        :param distances: the distances
-        :param known_distances: the known distances
-        :param known_distance_idx: the index of the known distances
-        :param unknown_distance_idx:  the index of the unknown distances
-        :param fluxes: the latent fluxes
[…]
