### 2021-06-26 0025244
Docstring for population_synth
