+++ popsynth/population_synth.py
@@ -1,4 +1,4 @@
-from abc import ABCMeta, abstractmethod
+from abc import ABCMeta
 from collections import OrderedDict
 from typing import Any, Dict, Optional, Union
 
@@ -12,15 +12,12 @@ from IPython.display import Markdown, Math, display
 
 # from numpy.typing import np.ndarray
 
-
 from popsynth.auxiliary_sampler import AuxiliarySampler, DerivedLumAuxSampler
 from popsynth.distribution import LuminosityDistribution, SpatialDistribution
 from popsynth.distributions.cosmological_distribution import CosmologicalDistribution
 from popsynth.population import Population
-from popsynth.selection_probability import ( SelectionProbabilty,
-                                             UnitySelection)
+from popsynth.selection_probability import SelectionProbabilty, UnitySelection
 from popsynth.utils.logging import setup_logger
-from popsynth.utils.progress_bar import progress_bar
 from popsynth.utils.registry import (
     auxiliary_parameter_registry,
     distribution_registry,
@@ -42,11 +39,11 @@ class PopulationSynth(object, metaclass=ABCMeta):
         derived luminosity distribution and everything is setup.
 
         :param spatial_distribution: The spatial distribution to sample locations from
-        :param luminosity_distribution: the optional luminosity distribution
-        :param seed: random see
-        :returns:
-        :rtype:
-
+        :type spatial_distribution: :class:`SpatialDistribution`
+        :param luminosity_distribution: The optional luminosity distribution
+        :type luminosity_distribution: :class:`LuminosityDistribution`
+        :param seed: Random seed
+        :type seed: int
         """
 
         self._n_model = 500  # type: int
@@ -85,13 +82,16 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
         self._has_derived_luminosity = False  # type: bool
         self._derived_luminosity_sampler = (
-            None)  # type: Union[DerivedLumAuxSampler, None]
+            None
+        )  # type: Union[DerivedLumAuxSampler, None]
 
         # set the selections be fully seen unless it is set by the user
         self._distance_selector: SelectionProbabilty = UnitySelection(
-            name="unity distance selector")
+            name="unity distance selector"
+        )
         self._flux_selector: SelectionProbabilty = UnitySelection(
-            name="unity flux selector")
+            name="unity flux selector"
+        )
 
         # check to see if the selectors are set
         self._distance_selector_set: bool = False
@@ -117,10 +117,11 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
     def clean(self, reset: bool = False):
         """
-        clean the auxiliary samplers, selections, etc
+        Clean the auxiliary samplers, selections, etc
         from the population synth
-        :param reset: reset any attached distributions and samplers
 
+        :param reset: If `True`, reset any attached distributions and samplers
+        :type reset: bool
         """
 
         if reset:
@@ -161,8 +162,7 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
             self._distance_selector.reset()
 
-        self._distance_selector = UnitySelection(
-            name="unity distance selector")
+        self._distance_selector = UnitySelection(name="unity distance selector")
 
         log.warning("removing spatial selector")
 
@@ -176,9 +176,10 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
     def write_to(self, file_name: str) -> None:
         """
-        write the population synth to a file
-        :param file_name: the file name of the output yaml
+        Write the population synth to a YAML file.
 
+        :param file_name: the file name of the output YAML
+        :type file_name: str
         """
         with open(file_name, "w") as f:
 
[…]
