### 2021-06-26 02531a6
Add GalacticPlaneSelection and update tests
