+++ popsynth/selection_probability/spatial_selection.py
@@ -1,6 +1,7 @@
-from popsynth.distribution import SpatialDistribution
+from astropy.coordinates import SkyCoord
 
-from .selection_probability import SelectionProbabilty
+from popsynth.distribution import SpatialDistribution
+from .selection_probability import SelectionProbabilty, SelectionParameter
 
 
 class SpatialSelection(SelectionProbabilty):
@@ -33,4 +34,36 @@ class SpatialSelection(SelectionProbabilty):
         self._spatial_distribution: SpatialDistribution = spatial_distribtuion
 
 
-__all__ = ["SpatialSelection"]
+class GalacticPlaneSelection(SpatialSelection):
+
+    _selection_name = "GalacticPlaneSelection"
+
+    b_limit = SelectionParameter(vmin=0, vmax=90)
+
+    def __init__(self, name: str = "galactic plane selector"):
+        """
+        A selection that excludes objects near the galactic plane.
+
+        :param name: Name of the selection
+        :type name: str
+        :param b_limit: Limit around Galactic plane to exclude in
+        Galactic latitude and in units of degrees
+        :type b_limit: :class:`SelectionParameter`
+        """
+        super(GalacticPlaneSelection, self).__init__(name=name)
+
+    def draw(self, size: int):
+
+        c = SkyCoord(
+            self._spatial_distribution.ra,
+            self._spatial_distribution.dec,
+            unit="deg",
+            frame="icrs",
+        )
+
+        b = c.galactic.b.deg
+
+        self._selection = (b >= self.b_limit) | (b <= -self.b_limit)
+
+
+__all__ = ["SpatialSelection", "GalacticPlaneSelection"]
