+++ popsynth/test/test_selections.py
@@ -1,38 +1,11 @@
-import numpy as np
-from astropy.coordinates import SkyCoord
-
 import popsynth
 
-
-class GalacticPlaneSelection(popsynth.SpatialSelection):
-
-    _selection_name = "GalacticPlaneSelection"
-
-    b_limit = popsynth.SelectionParameter(vmin=0, vmax=90)
-
-    def __init__(self, name="mw plane selector"):
-        """
-        places a limit above the galactic plane for objects
-        """
-        super(GalacticPlaneSelection, self).__init__(name=name)
-
-    def draw(self, size: int):
-
-        g_coor = SkyCoord(
-            self._spatial_distribution.ra,
-            self._spatial_distribution.dec,
-            unit="deg",
-            frame="icrs",
-        ).transform_to("galactic")
-
-        self._selection = (g_coor.b.deg >= self.b_limit) | (g_coor.b.deg <=
-                                                            -self.b_limit)
+from popsynth.selection_probability.spatial_selection import GalacticPlaneSelection
 
 
 def test_spatial_selection():
 
-    pg = popsynth.populations.Log10NormalHomogeneousSphericalPopulation(
-        10, 1, 1)
+    pg = popsynth.populations.Log10NormalHomogeneousSphericalPopulation(10, 1, 1)
 
     gps = GalacticPlaneSelection()
     gps.b_limit = 10.0
