+++ popsynth/utils/hdf5_utils.py
@@ -44,7 +44,9 @@ def recursively_save_dict_contents_to_group(h5file, path, dic: Dict[Any, Any]):
 
                 raise ValueError("Cannot save %s type" % type(item))
 
-        if isinstance(item, (np.ndarray, np.int64, np.float64, str, bytes, float, int)):
+        if isinstance(
+                item,
+            (np.ndarray, np.int64, np.float64, str, bytes, float, int)):
 
             h5file[path + "/" + key] = item
 
@@ -54,9 +56,9 @@ def recursively_save_dict_contents_to_group(h5file, path, dic: Dict[Any, Any]):
 
                 h5file[path + "/" + key] = "FILL_VALUE"
 
-            recursively_save_dict_contents_to_group(
-                h5file, path + "/" + key + "/", item
-            )
+            recursively_save_dict_contents_to_group(h5file,
+                                                    path + "/" + key + "/",
+                                                    item)
         else:
             raise ValueError("Cannot save %s type" % type(item))
 
@@ -91,8 +93,7 @@ def recursively_load_dict_contents_from_group(h5file, path):
 
         elif isinstance(item, h5py._hl.group.Group):
             ans[key] = recursively_load_dict_contents_from_group(
-                h5file, path + "/" + key + "/"
-            )
+                h5file, path + "/" + key + "/")
     return ans
 
 
