+++ popsynth/utils/configuration.py
@@ -11,7 +11,6 @@ _config_name = Path("popsynth_config.yml")
 
 _config_file = _config_path / _config_name
 
-
 # Define structure of configuration with dataclasses
 
 
@@ -60,7 +59,8 @@ if _config_file.is_file():
 
     _local_config = OmegaConf.load(_config_file)
 
-    popsynth_config: PopSynthConfig = OmegaConf.merge(popsynth_config, _local_config)
+    popsynth_config: PopSynthConfig = OmegaConf.merge(popsynth_config,
+                                                      _local_config)
 
 # Write defaults if not
 else:
