+++ popsynth/aux_samplers/normal_aux_sampler.py
@@ -34,15 +34,17 @@ class NormalAuxSampler(AuxiliarySampler):
 
     def true_sampler(self, size: int):
 
-        self._true_values = stats.norm.rvs(loc=self.mu, scale=self.tau, size=size)
+        self._true_values = stats.norm.rvs(loc=self.mu,
+                                           scale=self.tau,
+                                           size=size)
 
     def observation_sampler(self, size: int):
 
         if self._is_observed:
 
-            self._obs_values = stats.norm.rvs(
-                loc=self._true_values, scale=self.sigma, size=size
-            )
+            self._obs_values = stats.norm.rvs(loc=self._true_values,
+                                              scale=self.sigma,
+                                              size=size)
 
         else:
 
