+++ popsynth/distributions/log10_normal_distribution.py
@@ -26,15 +26,14 @@ class Log10NormalDistribution(LuminosityDistribution):
         """
         lf_form = r"\mathrm{Log_{10}Normal(\mu, \tau)}"
 
-        super(Log10NormalDistribution, self).__init__(
-            name=name, seed=seed, form=lf_form
-        )
+        super(Log10NormalDistribution, self).__init__(name=name,
+                                                      seed=seed,
+                                                      form=lf_form)
 
     def phi(self, L):
 
-        return (1.0 / (self.tau * L * np.sqrt(2 * np.pi))) * np.exp(
-            -((np.log10(L) - self.mu) ** 2) / (2 * self.tau ** 2)
-        )
+        return (1.0 / (self.tau * L * np.sqrt(2 * np.pi))) * np.exp(-(
+            (np.log10(L) - self.mu)**2) / (2 * self.tau**2))
 
     def draw_luminosity(self, size=1):
 
