+++ popsynth/aux_samplers/__init__.py
@@ -20,3 +20,4 @@ __all__ = [
     "ParetoAuxSampler",
     "PowerLawAuxSampler",
     "BrokenPowerLawAuxSampler",
+]
