### 2021-06-28 4dce6f2
Add skyplot to blazar example to show GalacticPlaneSelection
