### 2021-06-28 3784ee2
Tidy up and fix sphinx rendering of multi-line :param:s
