+++ popsynth/populations/schechter_populations.py
@@ -118,9 +118,9 @@ class SchechterZPowerCosmoPopulation(ZPowerCosmoPopulation):
         :param seed: Random seed
         :type seed: int
         :param is_rate: `True` if modelling a population of transient events,
-        `False` if modelling a population of steady-state objects.
-        Affects the ``time_adjustment`` method used in cosmo calculations.
-        Default is `True`.
+            `False` if modelling a population of steady-state objects.
+            Affects the ``time_adjustment`` method used in cosmo calculations.
+            Default is `True`.
         :type is_rate: bool
         """
 
@@ -177,9 +177,9 @@ class SchechterSFRPopulation(SFRPopulation):
         :param seed: Random seed
         :type seed: int
         :param is_rate: `True` if modelling a population of transient events,
-        `False` if modelling a population of steady-state objects.
-        Affects the ``time_adjustment`` method used in cosmo calculations.
-        Default is `True`.
+            `False` if modelling a population of steady-state objects.
+            Affects the ``time_adjustment`` method used in cosmo calculations.
+            Default is `True`.
         :type is_rate: bool
         """
 
