+++ popsynth/populations/lognormal_population.py
@@ -118,9 +118,9 @@ class LogNormalZPowerCosmoPopulation(ZPowerCosmoPopulation):
         :param seed: Random seed
         :type seed: int
         :param is_rate: `True` if modelling a population of transient events,
-        `False` if modelling a population of steady-state objects.
-        Affects the ``time_adjustment`` method used in cosmo calculations.
-        Default is `True`.
+            `False` if modelling a population of steady-state objects.
+            Affects the ``time_adjustment`` method used in cosmo calculations.
+            Default is `True`.
         :type is_rate: bool
         """
 
@@ -176,9 +176,9 @@ class LogNormalSFRPopulation(SFRPopulation):
         :param seed: Random seed
         :type seed: int
         :param is_rate: `True` if modelling a population of transient events,
-        `False` if modelling a population of steady-state objects.
-        Affects the ``time_adjustment`` method used in cosmo calculations.
-        Default is `True`.
+            `False` if modelling a population of steady-state objects.
+            Affects the ``time_adjustment`` method used in cosmo calculations.
+            Default is `True`.
         :type is_rate: bool
         """
 
@@ -308,9 +308,9 @@ class Log10NormalZPowerCosmoPopulation(ZPowerCosmoPopulation):
         :param seed: Random seed
         :type seed: int
         :param is_rate: `True` if modelling a population of transient events,
-        `False` if modelling a population of steady-state objects.
-        Affects the ``time_adjustment`` method used in cosmo calculations.
-        Default is `True`.
+            `False` if modelling a population of steady-state objects.
+            Affects the ``time_adjustment`` method used in cosmo calculations.
+            Default is `True`.
         :type is_rate: bool
         """
 
@@ -366,9 +366,9 @@ class Log10NormalSFRPopulation(SFRPopulation):
         :param seed: Random seed
         :type seed: int
         :param is_rate: `True` if modelling a population of transient events,
-        `False` if modelling a population of steady-state objects.
-        Affects the ``time_adjustment`` method used in cosmo calculations.
-        Default is `True`.
+            `False` if modelling a population of steady-state objects.
+            Affects the ``time_adjustment`` method used in cosmo calculations.
+            Default is `True`.
         :type is_rate: bool
         """
 
