+++ popsynth/utils/cosmology.py
@@ -1,5 +1,3 @@
-from dataclasses import dataclass
-
 import numpy as np
 from astropy.constants import c
 from astropy.cosmology import WMAP9 as cosmo
@@ -96,12 +94,14 @@ class Cosmology(object):
         """
         x = self.xx(z)
         z1 = 1.0 + z
-        val = ((2 * self.dh * z1 / self.Om_sqrt) *
-               (self.Phi(self.xx(0)) - 1.0 /
-                (np.sqrt(z1)) * self.Phi(x)) * 3.086e24)  # in cm
+        val = (
+            (2 * self.dh * z1 / self.Om_sqrt)
+            * (self.Phi(self.xx(0)) - 1.0 / (np.sqrt(z1)) * self.Phi(x))
+            * 3.086e24
+        )  # in cm
         return val
 
-    Def a(self, z):
+    def a(self, z):
 
         return np.sqrt(np.power(1 + z, 3.0) * self.Om + self.Ode)
 
