+++ popsynth/selection_probability/generic_selectors.py
@@ -46,7 +46,8 @@ class BernoulliSelection(SelectionProbabilty):
 
         if popsynth_config.show_progress:
 
-            self._selection = np.zeros(size, dtype=int).astype(bool)  # type: np.ndarray
+            self._selection = np.zeros(size, dtype=int).astype(
+                bool)  # type: np.ndarray
 
             for i in progress_bar(range(size), desc=f"Selecting {self.name}"):
 
@@ -57,9 +58,8 @@ class BernoulliSelection(SelectionProbabilty):
 
         else:
 
-            self._selection = stats.bernoulli.rvs(self.probability, size=size).astype(
-                bool
-            )  # type: np.ndarray
+            self._selection = stats.bernoulli.rvs(
+                self.probability, size=size).astype(bool)  # type: np.ndarray
 
 
 class BoxSelection(SelectionProbabilty):
@@ -317,17 +317,17 @@ class SoftSelection(SelectionProbabilty):
             values = self._observed_flux
 
         if not use_log:
-            probs = sf.expit(
-                self.strength * (values - self.boundary)
-            )  # type: np.ndarray
+            probs = sf.expit(self.strength *
+                             (values - self.boundary))  # type: np.ndarray
 
         else:
 
-            probs = sf.expit(
-                self.strength * (np.log10(values) - np.log10(self.boundary))
-            )  # type: np.ndarray
+            probs = sf.expit(self.strength *
+                             (np.log10(values) -
+                              np.log10(self.boundary)))  # type: np.ndarray
 
-        self._selection = stats.bernoulli.rvs(probs, size=len(values)).astype(bool)
+        self._selection = stats.bernoulli.rvs(probs,
+                                              size=len(values)).astype(bool)
 
     __all__ = [
         "UnitySelection",
