+++ popsynth/selection_probability/selection_probability.py
@@ -15,9 +15,9 @@ class SelectionParameter(Parameter):
     pass
 
 
-class SelectionProbabilty(
-    object, metaclass=AutoRegister(selection_registry, base_type=ParameterMeta)
-):
+class SelectionProbabilty(object,
+                          metaclass=AutoRegister(selection_registry,
+                                                 base_type=ParameterMeta)):
     def __init__(
         self,
         name: str = "name",
