+++ popsynth/population.py
@@ -24,7 +24,8 @@ from popsynth.utils.spherical_geometry import xyz
 log = setup_logger(__name__)
 
 wine = "#8F2727"
-dark, dark_highlight, mid, mid_highlight, light, light_highlight = betagen(wine)
+dark, dark_highlight, mid, mid_highlight, light, light_highlight = betagen(
+    wine)
 
 
 class Population(object):
@@ -118,8 +119,7 @@ class Population(object):
         self._phi = phi  # type: ArrayLike
 
         assert len(known_distances) + len(unknown_distance_idx) == sum(
-            selection
-        ), "the distances are not the correct size"
+            selection), "the distances are not the correct size"
 
         # latent fluxes
         self._fluxes = fluxes  # type: ArrayLike
@@ -472,7 +472,9 @@ class Population(object):
 
         for k, v in self._spatial_params.items():
 
-            spatial_grp.create_dataset(k, data=np.array([v]), compression="lzf")
+            spatial_grp.create_dataset(k,
+                                       data=np.array([v]),
+                                       compression="lzf")
 
         if self._lf_params is not None:
 
@@ -480,7 +482,9 @@ class Population(object):
 
             for k, v in self._lf_params.items():
 
-                lum_grp.create_dataset(k, data=np.array([v]), compression="lzf")
+                lum_grp.create_dataset(k,
+                                       data=np.array([v]),
+                                       compression="lzf")
 
             f.attrs["lf_form"] = np.string_(self._lf_form)
 
@@ -496,14 +500,16 @@ class Population(object):
         f.attrs["r_max"] = self._r_max
         f.attrs["seed"] = int(self._seed)
 
-        f.create_dataset("luminosities", data=self._luminosities, compression="lzf")
+        f.create_dataset("luminosities",
+                         data=self._luminosities,
+                         compression="lzf")
         f.create_dataset("distances", data=self._distances, compression="lzf")
-        f.create_dataset(
-            "known_distances", data=self._known_distances, compression="lzf"
-        )
-        f.create_dataset(
-            "known_distance_idx", data=self._known_distance_idx, compression="lzf"
-        )
+        f.create_dataset("known_distances",
+                         data=self._known_distances,
+                         compression="lzf")
+        f.create_dataset("known_distance_idx",
+                         data=self._known_distance_idx,
+                         compression="lzf")
         f.create_dataset(
             "unknown_distance_idx",
             data=self._unknown_distance_idx,
@@ -520,10 +526,12 @@ class Population(object):
         for k, v in self._auxiliary_quantities.items():
 
             q_grp = aux_grp.create_group(k)
-            q_grp.create_dataset(
-                "true_values", data=v["true_values"], compression="lzf"
-            )
-            q_grp.create_dataset("obs_values", data=v["obs_values"], compression="lzf")
+            q_grp.create_dataset("true_values",
+                                 data=v["true_values"],
+                                 compression="lzf")
+            q_grp.create_dataset("obs_values",
+                                 data=v["obs_values"],
+                                 compression="lzf")
 
         model_grp = f.create_group("model_spaces")
 
@@ -535,8 +543,7 @@ class Population(object):
         recursively_save_dict_contents_to_group(f, "truth", self._truth)
 
         recursively_save_dict_contents_to_group(
-            f, "graph", fill_graph_dict(nx.to_dict_of_dicts(self._graph))
-        )
+            f, "graph", fill_graph_dict(nx.to_dict_of_dicts(self._graph)))
 
         # now store the popsynth
         recursively_save_dict_contents_to_group(f, "popsynth", self._pop_synth)
@@ -653,8 +660,8 @@ class Population(object):
[…]
