### 2021-06-30 #3257 @grburgess
@warrickball Ok, Thanks to @cescalara , there is now another example in the docs as well as more complete doc strings. She has also now been added as a coauthor.
