### 2021-06-30 #3257 @HeloiseS
Hello!

Sorry for the delay. So much progress we're nearly there.

**Sidenote:** I am having issues with the website slowing my browser down significantly and freezing it - I think it might be the fancy 3D plots. I would recommend making the interactive plot an optional thing in a jupyter notebook somewhere else and instead having a static 3D plot maybe viewed from different angles to illustrate the point -> Or just warn people to firefox, instead of chrome if things get slow.

## Comments
* The fireflies can stay but the idea of making things accessible to PhD students isn't so much about making them cute, it is about not assuming knowledge and giving them the tools to learn more if some explanations are beyond the scope of your tutorial. For example, why is the luminosity function a power law - can we have a reference for the Pareto luminosity function? Put yourself in the shoes of someone who is new to the field - don't hesitate to cite classical references or even a textbook. Someone will be thankful for it.

* I really like that the level of detail you've got in most of your docstrings to describe the Luminosity functions - with the maths, and sometimes even the appropriate reference. I do have a question: how are people supposed to know which parameters for the distributions are physical and what isn't. Do they need to check the references or would it be easy for you to show or write somewhere the range of values that make sense?

* I love that you have an example of a transient population and an example of a non transient population - thanks to the other referee for suggesting that and to you for implementing it, it really makes all the difference. It's really cool to be able to see the underlying population and the observed one on top of each other.

* The docstrings need to be completed - remember this is the first point of contact for most coders: myclass.__doc__ or myclass()? are essential. I don't want to have to go back to the tutorials evertime I forget what a parameter does.
A bit more information would be nice on some of them, if you can ensure that they have, 3 things:
  - A description
  - The expected type and or format if relevant
  - Whether it is optional and if it is, the Default value

* I like the example of the selection bias towards massive stars! Could you add that to the docs I don't see it under "Examples"?

* Add community guidelines to the README on the github not just to the documentation
