+++ docs/md/short_grbs.md
@@ -1,0 +1,203 @@
+---
+jupyter:
+  jupytext:
+    formats: ipynb,md
+    text_representation:
+      extension: .md
+      format_name: markdown
+      format_version: '1.3'
+      jupytext_version: 1.11.1
+  kernelspec:
+    display_name: Python 3
+    language: python
+    name: python3
+---
+
+# Examples
+
+Here are some examples of astrophysical populations from publications.
+
+
+## Short GRBS 
+
+In [Ghirlanda et al. 2016](https://arxiv.org/abs/1607.07875) a fitting algorithm was used to determine the redshift and luminosity of short GRBS. We can use the parameters to reproduce the population and the observed GBM survey.
+
+```python
+from popsynth import SFRDistribution, BPLDistribution, PopulationSynth, NormalAuxSampler, AuxiliarySampler, HardFluxSelection
+from popsynth import update_logging_level
+update_logging_level("INFO")
+```
+
+```python
+%matplotlib inline
+
+import matplotlib.pyplot as plt
+from jupyterthemes import jtplot
+
+jtplot.style(context="notebook", fscale=1, grid=False)
+purple = "#B833FF"
+yellow = "#F6EF5B"
+
+import networkx as nx
+import numpy as np
+import warnings
+
+warnings.simplefilter("ignore")
+```
+
+In the work, the luminosity function of short GRBs is model as a broken power law.
+
+```python
+bpl = BPLDistribution()
+
+bpl.alpha = -0.53
+bpl.beta = -3.4
+bpl.Lmin = 1e47 # erg/s
+bpl.Lbreak = 2.8e52
+bpl.Lmax = 1e55
+
+```
+
+To model the redshift distribution, an empirical form from [Cole et al 2001](https://academic.oup.com/mnras/article/326/1/255/1026734?login=true) is used. In ```popsynth``` we call this the ```SFRDistribution``` (but perhaps a better name is needed).
+
+```python
+sfr = SFRDistribution()
+```
+
+```python
+sfr.r0 = 5.
+sfr.a = 1
+sfr.rise = 2.8
+sfr.decay = 3.5
+sfr.peak = 2.3
+```
+
+We can checkout how the rate changes with redshift
+
+```python
+fig, ax = plt.subplots()
+
+z = np.linspace(0,5,100)
+
+ax.plot(z, sfr.dNdV(z), color=purple)
+ax.set_xlabel("z")
+ax.set_ylabel(r"$\frac{\mathrm{d}N}{\mathrm{d}V}$")
+```
+
+<!-- #region -->
+In their model, the authors also have some secondary parameters that are connected to the luminosity. These are the  parameters for the spectrum of the GRB. It is proposed that the spectra peak energy (Ep) is linked to the luminosity by a power law relation:
+
+
+$$ \log E_{\mathrm{p}} \propto a + b \log L$$
+
+We can build an auxiliary sample to simulate this as well. But we will also add a bit of scatter to the intercept of the relation.
+<!-- #endregion -->
+
+```python
+intercept =NormalAuxSampler(name="intercept", observed=False)
+
+intercept.mu = 0.034
+intercept.sigma = .005
+
+
+
+```
+
+```python
+class EpSampler(AuxiliarySampler):
+    
+    _auxiliary_sampler_name = "EpSampler"
+
+    def __init__(self):
+
+        # pass up to the super class
+        super(EpSampler, self).__init__("Ep", observed=True, uses_luminosity = True)
+
+    def true_sampler(self, size):
+
+        # we will get the intercept's latent (true) value
+        # from its sampler
+        
+        intercept = self._secondary_samplers["intercept"].true_values
+        
+        slope = 0.84
+
[…]
