+++ docs/md/bl_lacs.md
@@ -1,0 +1,179 @@
+---
+jupyter:
+  jupytext:
+    formats: ipynb,md
+    text_representation:
+      extension: .md
+      format_name: markdown
+      format_version: '1.3'
+      jupytext_version: 1.11.1
+  kernelspec:
+    display_name: Python 3
+    language: python
+    name: python3
+---
+
+# BL Lac blazars
+A model for the luminosity function and cosmic evolution of BL Lac type blazars is presented in [Ajello et al. 2014](https://arxiv.org/abs/1310.0006), based on observations in gamma-rays with the Fermi-LAT instrument.
+
+We can use the results of this paper to build a BL Lac population that is able to reproduce the results reported in the recent [4FGL Fermi-LAT catalog](https://arxiv.org/abs/1902.10045) reasonably well.
+
+```python
+from scipy import special as sf
+from astropy.coordinates import SkyCoord
+from popsynth import (ZPowerCosmoDistribution, SoftFluxSelection,
+                      GalacticPlaneSelection)
+
+%matplotlib inline
+
+import matplotlib.pyplot as plt
+from jupyterthemes import jtplot
+
+jtplot.style(context="notebook", fscale=1, grid=False)
+purple = "#B833FF"
+yellow = "#F6EF5B"
+
+import networkx as nx
+import numpy as np
+import warnings
+
+warnings.simplefilter("ignore")
+
+```
+
+The work mentioned above presents 3 models for the BL Lac luminosity function. Here, we focus on the case of pure density evolution (referred to as PDE in the paper). In this case, the BL Lac population is parametrised as having a broken power law luminosity distribution, with an independent density evolution following a cosmological power law distribution.
+
+We work with a luminosity range of $L_\mathrm{min} = 7\times 10^{43}$ erg $\mathrm{s}^{-1}$ and $L_\mathrm{max} = 10^{52}$ erg $\mathrm{s}^{-1}$ following Ajello et al. 2014. All luminosities are in units of erg $\mathrm{s}^{-1}$. Similarly, the maxium redshift considered in $z=6$.
+
+We start by setting up the broken power law distribution (`BPLDistribution`).
+
+```python
+bpl = BPLDistribution()
+bpl.alpha = -1.5
+bpl.beta = -2.5
+bpl.Lmin = 7e43
+bpl.Lmax = 1e52
+bpl.Lbreak = 1e47
+
+fig, ax = plt.subplots()
+L = np.geomspace(bpl.Lmin, bpl.Lmax)
+ax.plot(L, bpl.phi(L), color=purple)
+ax.set_xscale("log")
+ax.set_yscale("log")
+ax.set_xlabel(r"L [erg $\mathrm{s}^{-1}$]")
+```
+
+We now move to the redshift distribution. Following the paper, we parametrize this as a negative power law in $z$. for the purpose of this example, we assume that Bl Lac blazars emit with a steady state. This means that we need to set the `is_rate` parameter to `False` when defining the `ZPowerCosmoDistribution` cosmological distribution. What we mean when we do this is that our local number density, `Lambda` is not per unit time. We also want to survey the whole sky, so we integrate over $4\pi$ sr in the value that we pass to the `Lambda`. 
+
+```python
+zpow = ZPowerCosmoDistribution()
+zpow.Lambda = 9000 # Gpc^-3 sr 
+zpow.delta = -6
+
+fig, ax = plt.subplots()
+z = np.linspace(0.01, 6)
+ax.plot(z, zpow.dNdV(z), color=purple)
+ax.set_yscale("log")
+ax.set_xlabel("z")
+ax.set_ylabel(r"$\frac{\mathrm{d}N}{\mathrm{d}V}$")
+```
+
[…]
