### 2021-07-02 f134ff7
rm old examples
