+++ docs/md/bl_lacs.md
@@ -23,6 +23,7 @@ from scipy import special as sf
 from astropy.coordinates import SkyCoord
 from popsynth import (ZPowerCosmoDistribution, SoftFluxSelection,
                       GalacticPlaneSelection)
+					  
 from popsynth import SFRDistribution, BPLDistribution, PopulationSynth, NormalAuxSampler, AuxiliarySampler, HardFluxSelection
 
 %matplotlib inline
