### 2021-07-08 88ac91d
adding better selection storage
