+++ popsynth/auxiliary_sampler.py
@@ -4,6 +4,7 @@ from typing import Any, Dict, List, Optional
 import numpy as np
 from class_registry import AutoRegister
 from numpy.typing import ArrayLike
+from dotmap import DotMap
 
 from popsynth.distribution import SpatialContainer
 from popsynth.selection_probability import SelectionProbabilty, UnitySelection
@@ -16,7 +17,130 @@ log = setup_logger(__name__)
 
 SamplerDict = Dict[str, Dict[str, ArrayLike]]
 
+class SecondaryContainer(object):
 
+    def __init__(self, name: str, true_values: ArrayLike, obs_values: ArrayLike, selection: ArrayLike) -> None:
+        """
+        A container for secondary properties that adds dict
+        and dictionary access
+
+        :param name: the name of the secondary
+        :type name: str
+        :param true_values: 
+        :type true_values: ArrayLike
+        :param obs_values: 
+        :type obs_values: ArrayLike
+        :param selection: 
+        :type selection: ArrayLike
+        :returns: 
+
+        """
+
+        self._true_values: ArrayLike = true_values
+        self._obs_values: ArrayLike = obs_values
+        self._selection: ArrayLike = selection
+
+        self._name: str = name
+
+    @property
+    def name(self) -> str:
+        return self._name
+        
+    @property
+    def true_values(self) -> ArrayLike:
+        """
+        The true (latent) values of the sampler
+
+        :returns: 
+
+        """
+        return self._true_values
+
+    @property
+    def obs_values(self) -> ArrayLike:
+        """
+        The observed values of the sampler
+
+        :returns: 
+
+        """
+        return self._obs_values
+
+    @property
+    def selection(self) -> ArrayLike:
+        """
+        The the slection of the values
+
+        :returns: 
+
+        """
+        return self._selection
+
+    def __getitem__(self, key):
+
+        if key == "selection":
+            return self._selection
+        
+        elif key == "true_values":
+            return self._true_values
+
+        elif key == "obs_values":
+            return self._obs_values
+        
+        else:
+
+            log.error("trying to access something that does not exist")
+            
+            raise RuntimeError()
+            
+
+
+class SecondaryStorage(DotMap):
+
+    def __init__(self):
+        """
+        A container for secondary samplers
+
+        :returns: 
+
+        """
+
+        super(SecondaryStorage, self).__init__()
+
+    def add_secondary(self, secondary_values: SecondaryContainer) -> None:
+        """
+        Add on a new secondary
+
+        :param secondary_values: 
+        :type secondary_values: SecondaryContainer
+        :returns: 
+
+        """
+        
+        self[secondary_values.name] = secondary_values
+        
+    def __add__(self, other):
+        
+        if self.empty():
+            return other
+
+        elif other.empty():
+            
+            return self
+        
+        else:
+            
+            for k, v in other.items():
+                
+                self[k] = v
+                
+            return self
+        
+        
+
+
+
+        
 class AuxiliaryParameter(Parameter):
     pass
 
@@ -94,9 +218,21 @@ class AuxiliarySampler(
 
     def set_selection_probability(self, selector: SelectionProbabilty) -> None:
 
-        assert isinstance(
+        """
+        Set a selection probabilty for this sampler. 
+
+        :param selector: A selection probability oobject
+        :type selector: SelectionProbabilty
+        :returns: 
+
+        """
+        if not isinstance(
             selector, SelectionProbabilty
-        ), "The selector is not a valid selection probability"
+        ):
+
+            log.error("The selector is not a valid selection probability")
+
+            raise AssertionError()
 
         self._selector = selector  # type: SelectionProbabilty
 
@@ -107,11 +243,21 @@ class AuxiliarySampler(
 
         self._selector.draw(len(self._obs_values))
 
[…]
