+++ setup.cfg
@@ -42,6 +42,7 @@ install_requires =
     numba
     pandas
     networkx
+    dotmap
     omegaconf
     colorama
     class-registry
