### 2021-07-08 87bfa6f
more doc strings
