### 2021-07-08 e207d82
better displays and repr s
