+++ popsynth/population_synth.py
@@ -1271,7 +1271,15 @@ class PopulationSynth(object, metaclass=ABCMeta):
         Display the simulation parameters.
         """
 
-        if self._luminosity_distribution is not None:
+
+        if self._has_derived_luminosity:
+
+            display(Markdown("## Luminosity Function"))
+
+            self._derived_luminosity_sampler.display()
+            
+        
+        elif self._luminosity_distribution is not None:
 
             display(Markdown("## Luminosity Function"))
 
@@ -1282,16 +1290,37 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
         self._spatial_distribution.display()
 
+        names = []
+
+        if self._has_derived_luminosity:
+
+            for k,v in self._derived_luminosity_sampler.secondary_samplers.items():
+
+                names.append(k)
+
+                display(Markdown(f"## {k}"))
+                
+                v.display()
+
+        
         for k,v in self._auxiliary_observations.items():
 
-            display(Markdown(f"## {k}"))
+            if k not in names:
 
-            v.display()
+                display(Markdown(f"## {k}"))
+
+                v.display()
 
         
     def __repr__(self) -> str:
 
-        if self._luminosity_distribution is not None:
+        if self._has_derived_luminosity:
+
+            out = "Luminosity Function\n"
+
+            out += self._derived_luminosity_sampler.__repr__()
+        
+        elif self._luminosity_distribution is not None:
         
             out = "Luminosity Function\n"
 
@@ -1302,9 +1331,21 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
         out += self._spatial_distribution.__repr__()
 
+        names = [] 
+        
+        if self._has_derived_luminosity:
+
+            for k,v in self._derived_luminosity_sampler.secondary_samplers.items():
+
+                names.append(k)
+
+                out += v.__repr__()
+        
         for k,v in self._auxiliary_observations.items():
 
-            out += v.__repr__()
+            if k not in names:
+
+                out += v.__repr__()
 
         
         return out 
