### 2021-07-09 #3257 @grburgess
Hi,

Sorry about the plots slowing things down. I have reduced the number of objects to try and make the plotting manageable.

Regarding the first two comments. Pareto is just a normalized power law. There are references to an economic principle, but the mathematical form should be sufficient as I would not want someone to think that it is in any way related to economics. As far as parameter ranges, this being a generic framework, there are no defined parameter ranges other than those that would make the functions mathematically incorrect (negative amplitudes for example). This is also why references really don't make much sense in this context as it is meant to be a frame work to build populations. The distributions are just normalized mathematical functions. One could place values of parameters that correspond to theoretical or observed values. Moreover, the description in the documentation for constructing custom distributions more important as other users have generally made their own distribution functions. As there is a lot of freedom, the only thing that can be documented is the construction.

The docstrings are now filled out.

I have added the massive stars example to the documentation.

I have added the community guidelines to the README.

@warrickball I have also now added text __repr__ so that quick information can be displayed in ipython
