+++ docs/index.rst
@@ -39,7 +39,6 @@ selection functions.
     notebooks/custom.ipynb
     notebooks/selections.ipynb
     notebooks/aux.ipynb
-    notebooks/milkyway.ipynb
     notebooks/contribute.ipynb
     api/API
 
