### 2021-07-10 #3257 @HeloiseS
okay cool just one question: is there a physical motivation for using this pareto function? Or is it a case of "we don't know what the luminosity function is doing for our populations so a powerlaw is a good guess"?
