### 2021-07-11 #3257 @grburgess
In general, luminosity functions are empirical descriptions that are modeling several underlying processes that manifest as a single observed distribution. For example, the GRB luminosity function is likely driven by progenitor mass / explosion energy, radiative process, and jet viewing angle. Observationally, this manifests as a broken power law. Thus, unless one is modeling all the physical processes, the most straight-forward modeling approach is to use some empirical function like a broken power law, Schecter function etc. On the simulation side, one can go any route they choose. If you want to model the luminosity distribution as an empirical function, then one does that. In `popsynth` the function of the ` DerivedLumAuxSampler` is that the value of luminosity perforce is a manifest property of underlying variables.

Nevertheless, there are several cases fo power laws being used to model luminosity functions. E.g., the radio luminosity function of of radio quiet AGN: https://iopscience.iop.org/article/10.1088/0004-637X/740/1/20 .

The inclusion of the Pareto distribution is simply because it is used a lot by those using this code, but as the software is provided as a framework to be extended, there is no preference for this distribution.
