+++ popsynth/auxiliary_sampler.py
@@ -9,9 +9,8 @@ import pandas as pd
 from IPython.display import Markdown, Math, display
 
 
-
 from popsynth.distribution import SpatialContainer
-from popsynth.selection_probability import SelectionProbabilty, UnitySelection
+from popsynth.selection_probability import SelectionProbability, UnitySelection
 from popsynth.utils.cosmology import cosmology
 from popsynth.utils.logging import setup_logger
 from popsynth.utils.meta import Parameter, ParameterMeta
@@ -23,21 +22,26 @@ SamplerDict = Dict[str, Dict[str, ArrayLike]]
 
 
 class SecondaryContainer(object):
-    def __init__(self, name: str, true_values: ArrayLike,
-                 obs_values: ArrayLike, selection: ArrayLike) -> None:
+    def __init__(
+        self,
+        name: str,
+        true_values: ArrayLike,
+        obs_values: ArrayLike,
+        selection: ArrayLike,
+    ) -> None:
         """
         A container for secondary properties that adds dict
         and dictionary access
 
         :param name: the name of the secondary
         :type name: str
-        :param true_values: 
+        :param true_values:
         :type true_values: ArrayLike
-        :param obs_values: 
+        :param obs_values:
         :type obs_values: ArrayLike
-        :param selection: 
+        :param selection:
         :type selection: ArrayLike
-        :returns: 
+        :returns:
 
         """
 
@@ -56,7 +60,7 @@ class SecondaryContainer(object):
         """
         The true (latent) values of the sampler
 
-        :returns: 
+        :returns:
 
         """
         return self._true_values
@@ -66,7 +70,7 @@ class SecondaryContainer(object):
         """
         The observed values of the sampler
 
-        :returns: 
+        :returns:
 
         """
         return self._obs_values
@@ -76,7 +80,7 @@ class SecondaryContainer(object):
         """
         The the slection of the values
 
-        :returns: 
+        :returns:
 
         """
         return self._selection
@@ -104,7 +108,7 @@ class SecondaryStorage(DotMap):
         """
         A container for secondary samplers
 
-        :returns: 
+        :returns:
 
         """
 
@@ -114,9 +118,9 @@ class SecondaryStorage(DotMap):
         """
         Add on a new secondary
 
-        :param secondary_values: 
+        :param secondary_values:
         :type secondary_values: SecondaryContainer
-        :returns: 
+        :returns:
 
         """
 
@@ -145,9 +149,8 @@ class AuxiliaryParameter(Parameter):
 
 
 class AuxiliarySampler(
-        object,
-        metaclass=AutoRegister(auxiliary_parameter_registry,
-                               base_type=ParameterMeta),
+    object,
+    metaclass=AutoRegister(auxiliary_parameter_registry, base_type=ParameterMeta),
 ):
     def __init__(
         self,
@@ -185,45 +188,42 @@ class AuxiliarySampler(
         self._parent_names = []
         self._has_secondary = False  # type: bool
         self._is_sampled = False  # type: bool
-        self._selector = UnitySelection()  # type: SelectionProbabilty
+        self._selector = UnitySelection()  # type: SelectionProbability
         self._uses_distance = uses_distance  # type: bool
         self._uses_luminosity = uses_luminosity  # type: bool
         self._uses_sky_position = uses_sky_position  # type: bool
 
-
     def display(self):
 
         out = {"parameter": [], "value": []}
-        
+
         for k, v in self._parameter_storage.items():
 
             out["parameter"].append(k)
             out["value"].append(v)
 
-
         display(pd.DataFrame(out))
-
 
     def __repr__(self):
 
         out = f"{self._name}\n"
-        
+
[…]
