+++ popsynth/population_synth.py
@@ -12,11 +12,16 @@ from IPython.display import Markdown, Math, display
 
 # from numpy.typing import np.ndarray
 
-from popsynth.auxiliary_sampler import AuxiliarySampler, DerivedLumAuxSampler, SecondaryContainer, SecondaryStorage
+from popsynth.auxiliary_sampler import (
+    AuxiliarySampler,
+    DerivedLumAuxSampler,
+    SecondaryContainer,
+    SecondaryStorage,
+)
 from popsynth.distribution import LuminosityDistribution, SpatialDistribution
 from popsynth.distributions.cosmological_distribution import CosmologicalDistribution
 from popsynth.population import Population
-from popsynth.selection_probability import SelectionProbabilty, UnitySelection
+from popsynth.selection_probability import SelectionProbability, UnitySelection
 from popsynth.utils.logging import setup_logger
 from popsynth.utils.registry import (
     auxiliary_parameter_registry,
@@ -82,18 +87,21 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
         self._has_derived_luminosity = False  # type: bool
         self._derived_luminosity_sampler = (
-            None)  # type: Union[DerivedLumAuxSampler, None]
+            None
+        )  # type: Union[DerivedLumAuxSampler, None]
 
         # set the selections be fully seen unless it is set by the user
-        self._distance_selector: SelectionProbabilty = UnitySelection(
-            name="unity distance selector")
-        self._flux_selector: SelectionProbabilty = UnitySelection(
-            name="unity flux selector")
+        self._distance_selector: SelectionProbability = UnitySelection(
+            name="unity distance selector"
+        )
+        self._flux_selector: SelectionProbability = UnitySelection(
+            name="unity flux selector"
+        )
 
         # check to see if the selectors are set
         self._distance_selector_set: bool = False
         self._flux_selector_set: bool = False
-        self._spatial_selector: Optional[SelectionProbabilty] = None
+        self._spatial_selector: Optional[SelectionProbability] = None
 
         self._params = {}  # type: dict
 
@@ -159,8 +167,7 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
             self._distance_selector.reset()
 
-        self._distance_selector = UnitySelection(
-            name="unity distance selector")
+        self._distance_selector = UnitySelection(name="unity distance selector")
 
         log.warning("removing spatial selector")
 
@@ -205,15 +212,14 @@ class PopulationSynth(object, metaclass=ABCMeta):
         spatial_distribution = {}
 
         spatial_distribution[
-            self._spatial_distribution.
-            _distribution_name] = self._spatial_distribution.truth
+            self._spatial_distribution._distribution_name
+        ] = self._spatial_distribution.truth
 
         # store is_rate if cosmological distribution
 
         if isinstance(self._spatial_distribution, CosmologicalDistribution):
 
-            spatial_distribution[
-                "is_rate"] = self._spatial_distribution._is_rate
+            spatial_distribution["is_rate"] = self._spatial_distribution._is_rate
 
         output["spatial distribution"] = spatial_distribution
 
@@ -225,8 +231,8 @@ class PopulationSynth(object, metaclass=ABCMeta):
             luminosity_distribution = {}
 
             luminosity_distribution[
-                self._luminosity_distribution.
-                _distribution_name] = self._luminosity_distribution.truth
+                self._luminosity_distribution._distribution_name
+            ] = self._luminosity_distribution.truth
 
             output["luminosity distribution"] = luminosity_distribution
 
@@ -234,8 +240,9 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
             flux_selection = {}
 
-            flux_selection[self._flux_selector.
-                           _selection_name] = self._flux_selector.parameters
[…]
