### 2021-07-14 a7fbc36
Add alpha=-1.0 case to plaw aux sampler
