### 2021-07-14 7d4e3cc
Add tests for plaw_aux_samplers
