### 2021-07-14 78cfc66
Fix default value for PowerLawAuxSampler.xmax
