+++ popsynth/auxiliary_sampler.py
@@ -8,7 +8,6 @@ from dotmap import DotMap
 import pandas as pd
 from IPython.display import Markdown, Math, display
 
-
 from popsynth.distribution import SpatialContainer
 from popsynth.selection_probability import SelectionProbability, UnitySelection
 from popsynth.utils.cosmology import cosmology
@@ -149,8 +148,9 @@ class AuxiliaryParameter(Parameter):
 
 
 class AuxiliarySampler(
-    object,
-    metaclass=AutoRegister(auxiliary_parameter_registry, base_type=ParameterMeta),
+        object,
+        metaclass=AutoRegister(auxiliary_parameter_registry,
+                               base_type=ParameterMeta),
 ):
     def __init__(
         self,
@@ -249,7 +249,8 @@ class AuxiliarySampler(
         self._dec = value.dec
         self._spatial_values = value
 
-    def set_selection_probability(self, selector: SelectionProbability) -> None:
+    def set_selection_probability(self,
+                                  selector: SelectionProbability) -> None:
         """
         Set a selection probabilty for this sampler.
 
@@ -362,12 +363,12 @@ class AuxiliarySampler(
             self._selector.set_observed_value(self._obs_values)
 
             # check to make sure we sampled!
-            assert (
-                self.true_values is not None and len(self.true_values) == size
-            ), f"{self.name} likely has a bad true_sampler function"
-            assert (
-                self.obs_values is not None and len(self.obs_values) == size
-            ), f"{self.name} likely has a observation_sampler function"
+            assert (self.true_values is not None
+                    and len(self.true_values) == size
+                    ), f"{self.name} likely has a bad true_sampler function"
+            assert (self.obs_values is not None
+                    and len(self.obs_values) == size
+                    ), f"{self.name} likely has a observation_sampler function"
 
             # now apply the selection to yourself
             # if there is nothing coded, it will be
@@ -394,7 +395,8 @@ class AuxiliarySampler(
 
         else:
 
-            log.debug(f"{self.name} is not reseting as it has not been sampled")
+            log.debug(
+                f"{self.name} is not reseting as it has not been sampled")
 
         for k, v in self._secondary_samplers.items():
 
@@ -449,16 +451,13 @@ class AuxiliarySampler(
                         self._graph.add_edge(spatial_distribution.name, k)
 
                 recursive_secondaries += v.get_secondary_properties(
-                    graph, k, spatial_distribution
-                )
+                    graph, k, spatial_distribution)
 
         # add our own on
 
         recursive_secondaries.add_secondary(
-            SecondaryContainer(
-                self._name, self._true_values, self._obs_values, self._selector
-            )
-        )
+            SecondaryContainer(self._name, self._true_values, self._obs_values,
+                               self._selector))
 
         return recursive_secondaries
 
@@ -483,8 +482,7 @@ class AuxiliarySampler(
 
             for k, v in self._secondary_samplers.items():
                 recursive_secondaries = v.get_secondary_objects(
-                    recursive_secondaries
-                )  # type: SamplerDict
+                    recursive_secondaries)  # type: SamplerDict
 
         # add our own on
 
@@ -675,9 +673,10 @@ class AuxiliarySampler(
 
 
 class NonObservedAuxSampler(AuxiliarySampler):
-    def __init__(
-        self, name: str, uses_distance: bool = False, uses_luminosity: bool = False
-    ):
+    def __init__(self,
+                 name: str,
+                 uses_distance: bool = False,
+                 uses_luminosity: bool = False):
 
         super(NonObservedAuxSampler, self).__init__(
             name=name,
@@ -698,9 +697,9 @@ class DerivedLumAuxSampler(AuxiliarySampler):
         :type uses_distance: bool
         """
 
[…]
