+++ popsynth/population_synth.py
@@ -87,16 +87,13 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
         self._has_derived_luminosity = False  # type: bool
         self._derived_luminosity_sampler = (
-            None
-        )  # type: Union[DerivedLumAuxSampler, None]
+            None)  # type: Union[DerivedLumAuxSampler, None]
 
         # set the selections be fully seen unless it is set by the user
         self._distance_selector: SelectionProbability = UnitySelection(
-            name="unity distance selector"
-        )
+            name="unity distance selector")
         self._flux_selector: SelectionProbability = UnitySelection(
-            name="unity flux selector"
-        )
+            name="unity flux selector")
 
         # check to see if the selectors are set
         self._distance_selector_set: bool = False
@@ -167,7 +164,8 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
             self._distance_selector.reset()
 
-        self._distance_selector = UnitySelection(name="unity distance selector")
+        self._distance_selector = UnitySelection(
+            name="unity distance selector")
 
         log.warning("removing spatial selector")
 
@@ -212,14 +210,15 @@ class PopulationSynth(object, metaclass=ABCMeta):
         spatial_distribution = {}
 
         spatial_distribution[
-            self._spatial_distribution._distribution_name
-        ] = self._spatial_distribution.truth
+            self._spatial_distribution.
+            _distribution_name] = self._spatial_distribution.truth
 
         # store is_rate if cosmological distribution
 
         if isinstance(self._spatial_distribution, CosmologicalDistribution):
 
-            spatial_distribution["is_rate"] = self._spatial_distribution._is_rate
+            spatial_distribution[
+                "is_rate"] = self._spatial_distribution._is_rate
 
         output["spatial distribution"] = spatial_distribution
 
@@ -231,8 +230,8 @@ class PopulationSynth(object, metaclass=ABCMeta):
             luminosity_distribution = {}
 
             luminosity_distribution[
-                self._luminosity_distribution._distribution_name
-            ] = self._luminosity_distribution.truth
+                self._luminosity_distribution.
+                _distribution_name] = self._luminosity_distribution.truth
 
             output["luminosity distribution"] = luminosity_distribution
 
@@ -240,9 +239,8 @@ class PopulationSynth(object, metaclass=ABCMeta):
 
             flux_selection = {}
 
-            flux_selection[
-                self._flux_selector._selection_name
-            ] = self._flux_selector.parameters
+            flux_selection[self._flux_selector.
+                           _selection_name] = self._flux_selector.parameters
 
             output["flux selection"] = flux_selection
 
@@ -251,8 +249,8 @@ class PopulationSynth(object, metaclass=ABCMeta):
             distance_selection = {}
 
             distance_selection[
-                self._distance_selector._selection_name
-            ] = self._distance_selector.parameters
+                self._distance_selector.
+                _selection_name] = self._distance_selector.parameters
 
             output["distance selection"] = distance_selection
 
@@ -261,8 +259,8 @@ class PopulationSynth(object, metaclass=ABCMeta):
             spatial_selection = {}
 
             spatial_selection[
-                self._spatial_selector._selection_name
-            ] = self._spatial_selector.parameters
+                self._spatial_selector.
+                _selection_name] = self._spatial_selector.parameters
 
             output["spatial selection"] = spatial_selection
 
@@ -320,8 +318,7 @@ class PopulationSynth(object, metaclass=ABCMeta):
             # create the instance
 
             luminosity_distribution: LuminosityDistribution = distribution_registry[
[…]
