### 2021-07-15 #3257 @warrickball
I just had a look at the latest PDF draft and have a two very minor comments.

First, the opening sentence,

> Simulating a population survey of fluxes and redshifts (distances) from an astrophysical population is a routine task.

reads slightly strangely to me, I think because I don't quite know what a "population survey ... from an astrophysical population" is. It makes more sense to me if the first occurrence of "population" is deleted.

Second, something has gone wrong with this reference:

> al., A. A. et. (2021). SkyPy: A package for modelling the universe. In GitHub repository. GitHub.https://github.com/skypyproject/skypy

The only other thing I'd like to check again is the state of the automated tests. I can see that the CI appears to be working but I just want to check that everything makes sense for potential contributors. Since @HeloiseS is happy, I'll try to do this today but I might have to work on some more urgent things that'll push this back to late next week.
