+++ paper.bib
@@ -1,5 +1,5 @@
 @Misc{skypy,
-  author = {A. Amara et al.},
+  author = {A. Amara and et al.},
   title = {SkyPy: A package for modelling the Universe.},
   year = {2021},
   publisher = {GitHub},
