### 2021-07-15 #3257 @grburgess
@warrickball yes, there are too many populations in that sentence. I removed the first one.

I see that SkyPy has a joss-paper under review, so I grabbed the first author from there.
