### 2021-07-15 #3257 @warrickball
I updated a previous checkout and was able to run the tests successfully by running `pytest` in the `popsynth` folder. It now looks like this isn't documented but should be, probably in a section near *Contributing*.

I also noticed that `test_schecter` (sic) takes a long time. I didn't record any timing information but it was many minutes. The CI report for the Ubuntu build of the [most recent commit](https://github.com/grburgess/popsynth/runs/3074526020) shows the tests taking 16 minutes. You might be happy with that but it seems surprisingly long and probably dominated by that one test. In my experience, if tests take a long time, people won't run them or won't run them often enough to catch bugs.

The tests do run, though, so my checklist is complete and I'm happy for this to be accepted in JOSS.
