### 2021-07-16 #3257 @xuanxu
Thanks @HeloiseS & @warrickball for your reviews!

@grburgess I think I spotted [a typo](https://github.com/grburgess/popsynth/pull/44) in the paper' summary
