### 2021-07-16 #3257 @xuanxu
OK @grburgess, everything looks good, here are the next steps:

- Please release a new tagged version from the current master branch so it includes all the changes made during the review process
- Then archive that latest release in Zenodo or a similar service
- Check the Zenodo deposit has the correct metadata: title and authors names should match the paper; you may also add authors' ORCIDs.

Once you do that please report here the version number and archive DOI
