### 2021-07-16 #3257 @grburgess
@xuanxu I usually do a tag which then releases to pypi. should I just tar the pypi release and put that in zenodo ?
