### 2021-07-16 #3257 @xuanxu
@grburgess yes, you can upload the .zip or the .tar file generated by GH when the tag (or release) is created.

![files](https://user-images.githubusercontent.com/6528/125932763-15554a61-bc50-485a-a755-ef98cd2fe4a2.png)
