### 2021-07-16 #3257 @grburgess
it is also on pypi: https://pypi.org/project/popsynth
