+++ README.md
@@ -3,7 +3,7 @@
 [![codecov](https://codecov.io/gh/grburgess/popsynth/branch/master/graph/badge.svg)](https://codecov.io/gh/grburgess/popsynth)
 [![Codacy Badge](https://api.codacy.com/project/badge/Grade/5d02c9e6f5c540989a615eb1575863e3)](https://app.codacy.com/gh/grburgess/popsynth?utm_source=github.com&utm_medium=referral&utm_content=grburgess/popsynth&utm_campaign=Badge_Grade_Settings)
 [![Documentation Status](https://readthedocs.org/projects/popsynth/badge/?version=latest)](https://popsynth.readthedocs.io/en/latest/?badge=latest)
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3372456.svg)](https://doi.org/10.5281/zenodo.3372456)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5109583.svg)](https://doi.org/10.5281/zenodo.5109583)
 ![PyPI](https://img.shields.io/pypi/v/popsynth)
 ![PyPI - Downloads](https://img.shields.io/pypi/dm/popsynth)
 # popsynth
