### 2021-07-16 #3257 @xuanxu
@grburgess Yes, the Zenodo snapshot's DOI will be archived associated to the paper, so it needs to match title and authors. You can change the title and other metadata directly in the Zenodo UI (Edit--> Basic information --> title), there's no need to update the files. I suggest you get a new DOI, with all correct data (right now, the publication date appears as Aug 2019 and the link to the repo in the sidebar point to the 0.2.2 tree)
