### 2021-07-16 #3257 @xuanxu
@grburgess Please upload to Zenodo the [GitHub generated](https://github.com/grburgess/popsynth/releases/tag/v1.0.3) [zip file](https://github.com/grburgess/popsynth/archive/refs/tags/v1.0.3.zip) instead of the tar pushed to pypi –this tar file is a distribution file where many of the repo's contents are not included, like docs, tests, etc.  The tag's zip/tar.gz is an actual snapshot of the repository.
(sorry for the back and forth)
